Add specific database query builder classes to rdbms
