+++ tests/Database/DatabaseQueryBuilderTest.php
@@ -4,18 +4,11 @@ namespace Illuminate\Tests\Database;
 
 use BadMethodCallException;
 use Closure;
-use DateTime;
 use Illuminate\Database\ConnectionInterface;
 use Illuminate\Database\Eloquent\Builder as EloquentBuilder;
 use Illuminate\Database\Query\Builder;
 use Illuminate\Database\Query\Expression as Raw;
 use Illuminate\Database\Query\Grammars\Grammar;
-use Illuminate\Database\Query\Grammars\MySqlGrammar;
-use Illuminate\Database\Query\Grammars\PostgresGrammar;
-use Illuminate\Database\Query\Grammars\SQLiteGrammar;
-use Illuminate\Database\Query\Grammars\SqlServerGrammar;
-use Illuminate\Database\Query\Processors\MySqlProcessor;
-use Illuminate\Database\Query\Processors\PostgresProcessor;
 use Illuminate\Database\Query\Processors\Processor;
 use Illuminate\Pagination\AbstractPaginator as Paginator;
 use Illuminate\Pagination\Cursor;
@@ -68,19 +61,6 @@ class DatabaseQueryBuilderTest extends TestCase
         $this->assertNull($builder->columns);
     }
 
-    public function testBasicSelectUseWritePdo()
-    {
-        $builder = $this->getMySqlBuilderWithProcessor();
-        $builder->getConnection()->shouldReceive('select')->once()
-            ->with('select * from `users`', [], false);
-        $builder->useWritePdo()->select('*')->from('users')->get();
-
-        $builder = $this->getMySqlBuilderWithProcessor();
-        $builder->getConnection()->shouldReceive('select')->once()
-            ->with('select * from `users`', [], true);
-        $builder->select('*')->from('users')->get();
-    }
-
     public function testBasicTableWrappingProtectsQuotationMarks()
     {
         $builder = $this->getBuilder();
@@ -129,10 +109,6 @@ class DatabaseQueryBuilderTest extends TestCase
         $builder = $this->getBuilder();
         $builder->distinct('foo')->select('foo', 'bar')->from('users');
         $this->assertSame('select distinct "foo", "bar" from "users"', $builder->toSql());
-
-        $builder = $this->getPostgresBuilder();
-        $builder->distinct('foo')->select('foo', 'bar')->from('users');
-        $this->assertSame('select distinct on ("foo") "foo", "bar" from "users"', $builder->toSql());
     }
 
     public function testBasicAlias()
@@ -338,51 +314,6 @@ class DatabaseQueryBuilderTest extends TestCase
         $this->assertEquals([0 => 12], $builder->getBindings());
     }
 
-    public function testMySqlWrappingProtectsQuotationMarks()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->From('some`table');
-        $this->assertSame('select * from `some``table`', $builder->toSql());
-    }
-
-    public function testDateBasedWheresAcceptsTwoArguments()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->whereDate('created_at', 1);
-        $this->assertSame('select * from `users` where date(`created_at`) = ?', $builder->toSql());
-
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->whereDay('created_at', 1);
-        $this->assertSame('select * from `users` where day(`created_at`) = ?', $builder->toSql());
-
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->whereMonth('created_at', 1);
-        $this->assertSame('select * from `users` where month(`created_at`) = ?', $builder->toSql());
-
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->whereYear('created_at', 1);
-        $this->assertSame('select * from `users` where year(`created_at`) = ?', $builder->toSql());
-    }
-
-    public function testDateBasedOrWheresAcceptsTwoArguments()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->where('id', 1)->orWhereDate('created_at', 1);
-        $this->assertSame('select * from `users` where `id` = ? or date(`created_at`) = ?', $builder->toSql());
-
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->where('id', 1)->orWhereDay('created_at', 1);
-        $this->assertSame('select * from `users` where `id` = ? or day(`created_at`) = ?', $builder->toSql());
-
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->where('id', 1)->orWhereMonth('created_at', 1);
-        $this->assertSame('select * from `users` where `id` = ? or month(`created_at`) = ?', $builder->toSql());
-
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->where('id', 1)->orWhereYear('created_at', 1);
-        $this->assertSame('select * from `users` where `id` = ? or year(`created_at`) = ?', $builder->toSql());
-    }
-
     public function testDateBasedWheresExpressionIsNotBound()
     {
         $builder = $this->getBuilder();
@@ -400,263 +331,6 @@ class DatabaseQueryBuilderTest extends TestCase
         $builder = $this->getBuilder();
         $builder->select('*')->from('users')->whereYear('created_at', new Raw('NOW()'));
         $this->assertEquals([], $builder->getBindings());
-    }
-
-    public function testWhereDateMySql()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->whereDate('created_at', '=', '2015-12-21');
-        $this->assertSame('select * from `users` where date(`created_at`) = ?', $builder->toSql());
-        $this->assertEquals([0 => '2015-12-21'], $builder->getBindings());
-
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->whereDate('created_at', '=', new Raw('NOW()'));
-        $this->assertSame('select * from `users` where date(`created_at`) = NOW()', $builder->toSql());
-    }
-
-    public function testWhereDayMySql()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->whereDay('created_at', '=', 1);
-        $this->assertSame('select * from `users` where day(`created_at`) = ?', $builder->toSql());
-        $this->assertEquals([0 => 1], $builder->getBindings());
-    }
-
-    public function testOrWhereDayMySql()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->whereDay('created_at', '=', 1)->orWhereDay('created_at', '=', 2);
-        $this->assertSame('select * from `users` where day(`created_at`) = ? or day(`created_at`) = ?', $builder->toSql());
-        $this->assertEquals([0 => 1, 1 => 2], $builder->getBindings());
-    }
-
-    public function testWhereMonthMySql()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->whereMonth('created_at', '=', 5);
-        $this->assertSame('select * from `users` where month(`created_at`) = ?', $builder->toSql());
-        $this->assertEquals([0 => 5], $builder->getBindings());
-    }
-
-    public function testOrWhereMonthMySql()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->whereMonth('created_at', '=', 5)->orWhereMonth('created_at', '=', 6);
-        $this->assertSame('select * from `users` where month(`created_at`) = ? or month(`created_at`) = ?', $builder->toSql());
-        $this->assertEquals([0 => 5, 1 => 6], $builder->getBindings());
-    }
-
-    public function testWhereYearMySql()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->whereYear('created_at', '=', 2014);
-        $this->assertSame('select * from `users` where year(`created_at`) = ?', $builder->toSql());
-        $this->assertEquals([0 => 2014], $builder->getBindings());
-    }
-
-    public function testOrWhereYearMySql()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->whereYear('created_at', '=', 2014)->orWhereYear('created_at', '=', 2015);
-        $this->assertSame('select * from `users` where year(`created_at`) = ? or year(`created_at`) = ?', $builder->toSql());
-        $this->assertEquals([0 => 2014, 1 => 2015], $builder->getBindings());
-    }
-
-    public function testWhereTimeMySql()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->whereTime('created_at', '>=', '22:00');
-        $this->assertSame('select * from `users` where time(`created_at`) >= ?', $builder->toSql());
-        $this->assertEquals([0 => '22:00'], $builder->getBindings());
-    }
-
-    public function testWhereTimeOperatorOptionalMySql()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->whereTime('created_at', '22:00');
-        $this->assertSame('select * from `users` where time(`created_at`) = ?', $builder->toSql());
-        $this->assertEquals([0 => '22:00'], $builder->getBindings());
-    }
-
-    public function testWhereTimeOperatorOptionalPostgres()
-    {
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->whereTime('created_at', '22:00');
-        $this->assertSame('select * from "users" where "created_at"::time = ?', $builder->toSql());
-        $this->assertEquals([0 => '22:00'], $builder->getBindings());
-    }
-
-    public function testWhereTimeSqlServer()
-    {
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->whereTime('created_at', '22:00');
-        $this->assertSame('select * from [users] where cast([created_at] as time) = ?', $builder->toSql());
-        $this->assertEquals([0 => '22:00'], $builder->getBindings());
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->whereTime('created_at', new Raw('NOW()'));
-        $this->assertSame('select * from [users] where cast([created_at] as time) = NOW()', $builder->toSql());
-        $this->assertEquals([], $builder->getBindings());
-    }
-
-    public function testWhereDatePostgres()
-    {
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->whereDate('created_at', '=', '2015-12-21');
-        $this->assertSame('select * from "users" where "created_at"::date = ?', $builder->toSql());
-        $this->assertEquals([0 => '2015-12-21'], $builder->getBindings());
-
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->whereDate('created_at', new Raw('NOW()'));
-        $this->assertSame('select * from "users" where "created_at"::date = NOW()', $builder->toSql());
-    }
-
-    public function testWhereDayPostgres()
-    {
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->whereDay('created_at', '=', 1);
-        $this->assertSame('select * from "users" where extract(day from "created_at") = ?', $builder->toSql());
-        $this->assertEquals([0 => 1], $builder->getBindings());
-    }
-
-    public function testWhereMonthPostgres()
-    {
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->whereMonth('created_at', '=', 5);
-        $this->assertSame('select * from "users" where extract(month from "created_at") = ?', $builder->toSql());
-        $this->assertEquals([0 => 5], $builder->getBindings());
-    }
-
-    public function testWhereYearPostgres()
-    {
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->whereYear('created_at', '=', 2014);
-        $this->assertSame('select * from "users" where extract(year from "created_at") = ?', $builder->toSql());
-        $this->assertEquals([0 => 2014], $builder->getBindings());
-    }
-
-    public function testWhereTimePostgres()
-    {
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->whereTime('created_at', '>=', '22:00');
-        $this->assertSame('select * from "users" where "created_at"::time >= ?', $builder->toSql());
-        $this->assertEquals([0 => '22:00'], $builder->getBindings());
-    }
-
-    public function testWhereLikePostgres()
-    {
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->where('id', 'like', '1');
-        $this->assertSame('select * from "users" where "id"::text like ?', $builder->toSql());
-        $this->assertEquals([0 => '1'], $builder->getBindings());
-
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->where('id', 'LIKE', '1');
-        $this->assertSame('select * from "users" where "id"::text LIKE ?', $builder->toSql());
-        $this->assertEquals([0 => '1'], $builder->getBindings());
-
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->where('id', 'ilike', '1');
-        $this->assertSame('select * from "users" where "id"::text ilike ?', $builder->toSql());
-        $this->assertEquals([0 => '1'], $builder->getBindings());
-
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->where('id', 'not like', '1');
-        $this->assertSame('select * from "users" where "id"::text not like ?', $builder->toSql());
-        $this->assertEquals([0 => '1'], $builder->getBindings());
-
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->where('id', 'not ilike', '1');
-        $this->assertSame('select * from "users" where "id"::text not ilike ?', $builder->toSql());
-        $this->assertEquals([0 => '1'], $builder->getBindings());
-    }
-
-    public function testWhereDateSqlite()
-    {
-        $builder = $this->getSQLiteBuilder();
-        $builder->select('*')->from('users')->whereDate('created_at', '=', '2015-12-21');
-        $this->assertSame('select * from "users" where strftime(\'%Y-%m-%d\', "created_at") = cast(? as text)', $builder->toSql());
-        $this->assertEquals([0 => '2015-12-21'], $builder->getBindings());
-
-        $builder = $this->getSQLiteBuilder();
-        $builder->select('*')->from('users')->whereDate('created_at', new Raw('NOW()'));
-        $this->assertSame('select * from "users" where strftime(\'%Y-%m-%d\', "created_at") = cast(NOW() as text)', $builder->toSql());
-    }
-
-    public function testWhereDaySqlite()
-    {
-        $builder = $this->getSQLiteBuilder();
-        $builder->select('*')->from('users')->whereDay('created_at', '=', 1);
-        $this->assertSame('select * from "users" where strftime(\'%d\', "created_at") = cast(? as text)', $builder->toSql());
-        $this->assertEquals([0 => 1], $builder->getBindings());
-    }
-
-    public function testWhereMonthSqlite()
-    {
-        $builder = $this->getSQLiteBuilder();
-        $builder->select('*')->from('users')->whereMonth('created_at', '=', 5);
-        $this->assertSame('select * from "users" where strftime(\'%m\', "created_at") = cast(? as text)', $builder->toSql());
-        $this->assertEquals([0 => 5], $builder->getBindings());
-    }
-
-    public function testWhereYearSqlite()
-    {
-        $builder = $this->getSQLiteBuilder();
-        $builder->select('*')->from('users')->whereYear('created_at', '=', 2014);
-        $this->assertSame('select * from "users" where strftime(\'%Y\', "created_at") = cast(? as text)', $builder->toSql());
-        $this->assertEquals([0 => 2014], $builder->getBindings());
-    }
-
-    public function testWhereTimeSqlite()
-    {
-        $builder = $this->getSQLiteBuilder();
-        $builder->select('*')->from('users')->whereTime('created_at', '>=', '22:00');
-        $this->assertSame('select * from "users" where strftime(\'%H:%M:%S\', "created_at") >= cast(? as text)', $builder->toSql());
-        $this->assertEquals([0 => '22:00'], $builder->getBindings());
-    }
-
-    public function testWhereTimeOperatorOptionalSqlite()
-    {
-        $builder = $this->getSQLiteBuilder();
-        $builder->select('*')->from('users')->whereTime('created_at', '22:00');
-        $this->assertSame('select * from "users" where strftime(\'%H:%M:%S\', "created_at") = cast(? as text)', $builder->toSql());
-        $this->assertEquals([0 => '22:00'], $builder->getBindings());
-    }
-
-    public function testWhereDateSqlServer()
-    {
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->whereDate('created_at', '=', '2015-12-21');
-        $this->assertSame('select * from [users] where cast([created_at] as date) = ?', $builder->toSql());
-        $this->assertEquals([0 => '2015-12-21'], $builder->getBindings());
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->whereDate('created_at', new Raw('NOW()'));
-        $this->assertSame('select * from [users] where cast([created_at] as date) = NOW()', $builder->toSql());
-    }
-
-    public function testWhereDaySqlServer()
-    {
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->whereDay('created_at', '=', 1);
-        $this->assertSame('select * from [users] where day([created_at]) = ?', $builder->toSql());
-        $this->assertEquals([0 => 1], $builder->getBindings());
-    }
-
-    public function testWhereMonthSqlServer()
-    {
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->whereMonth('created_at', '=', 5);
-        $this->assertSame('select * from [users] where month([created_at]) = ?', $builder->toSql());
-        $this->assertEquals([0 => 5], $builder->getBindings());
-    }
-
-    public function testWhereYearSqlServer()
-    {
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->whereYear('created_at', '=', 2014);
-        $this->assertSame('select * from [users] where year([created_at]) = ?', $builder->toSql());
-        $this->assertEquals([0 => 2014], $builder->getBindings());
     }
 
     public function testWhereBetweens()
@@ -886,112 +560,12 @@ class DatabaseQueryBuilderTest extends TestCase
         $this->assertEquals([], $builder->getBindings());
     }
 
-    public function testWhereFulltextMySql()
-    {
-        $builder = $this->getMySqlBuilderWithProcessor();
-        $builder->select('*')->from('users')->whereFulltext('body', 'Hello World');
-        $this->assertSame('select * from `users` where match (`body`) against (? in natural language mode)', $builder->toSql());
-        $this->assertEquals(['Hello World'], $builder->getBindings());
-
-        $builder = $this->getMySqlBuilderWithProcessor();
-        $builder->select('*')->from('users')->whereFulltext('body', 'Hello World', ['expanded' => true]);
-        $this->assertSame('select * from `users` where match (`body`) against (? in natural language mode with query expansion)', $builder->toSql());
-        $this->assertEquals(['Hello World'], $builder->getBindings());
-
-        $builder = $this->getMySqlBuilderWithProcessor();
-        $builder->select('*')->from('users')->whereFulltext('body', '+Hello -World', ['mode' => 'boolean']);
-        $this->assertSame('select * from `users` where match (`body`) against (? in boolean mode)', $builder->toSql());
-        $this->assertEquals(['+Hello -World'], $builder->getBindings());
-
-        $builder = $this->getMySqlBuilderWithProcessor();
-        $builder->select('*')->from('users')->whereFulltext('body', '+Hello -World', ['mode' => 'boolean', 'expanded' => true]);
-        $this->assertSame('select * from `users` where match (`body`) against (? in boolean mode)', $builder->toSql());
-        $this->assertEquals(['+Hello -World'], $builder->getBindings());
-
-        $builder = $this->getMySqlBuilderWithProcessor();
-        $builder->select('*')->from('users')->whereFulltext(['body', 'title'], 'Car,Plane');
-        $this->assertSame('select * from `users` where match (`body`, `title`) against (? in natural language mode)', $builder->toSql());
-        $this->assertEquals(['Car,Plane'], $builder->getBindings());
-    }
-
-    public function testWhereFulltextPostgres()
-    {
-        $builder = $this->getPostgresBuilderWithProcessor();
-        $builder->select('*')->from('users')->whereFulltext('body', 'Hello World');
-        $this->assertSame('select * from "users" where (to_tsvector(\'english\', "body")) @@ plainto_tsquery(\'english\', ?)', $builder->toSql());
-        $this->assertEquals(['Hello World'], $builder->getBindings());
-
-        $builder = $this->getPostgresBuilderWithProcessor();
-        $builder->select('*')->from('users')->whereFulltext('body', 'Hello World', ['language' => 'simple']);
-        $this->assertSame('select * from "users" where (to_tsvector(\'simple\', "body")) @@ plainto_tsquery(\'simple\', ?)', $builder->toSql());
-        $this->assertEquals(['Hello World'], $builder->getBindings());
-
-        $builder = $this->getPostgresBuilderWithProcessor();
-        $builder->select('*')->from('users')->whereFulltext('body', 'Hello World', ['mode' => 'plain']);
-        $this->assertSame('select * from "users" where (to_tsvector(\'english\', "body")) @@ plainto_tsquery(\'english\', ?)', $builder->toSql());
-        $this->assertEquals(['Hello World'], $builder->getBindings());
-
-        $builder = $this->getPostgresBuilderWithProcessor();
-        $builder->select('*')->from('users')->whereFulltext('body', 'Hello World', ['mode' => 'phrase']);
-        $this->assertSame('select * from "users" where (to_tsvector(\'english\', "body")) @@ phraseto_tsquery(\'english\', ?)', $builder->toSql());
-        $this->assertEquals(['Hello World'], $builder->getBindings());
-
-        $builder = $this->getPostgresBuilderWithProcessor();
-        $builder->select('*')->from('users')->whereFulltext('body', '+Hello -World', ['mode' => 'websearch']);
-        $this->assertSame('select * from "users" where (to_tsvector(\'english\', "body")) @@ websearch_to_tsquery(\'english\', ?)', $builder->toSql());
-        $this->assertEquals(['+Hello -World'], $builder->getBindings());
-
-        $builder = $this->getPostgresBuilderWithProcessor();
-        $builder->select('*')->from('users')->whereFulltext('body', 'Hello World', ['language' => 'simple', 'mode' => 'plain']);
-        $this->assertSame('select * from "users" where (to_tsvector(\'simple\', "body")) @@ plainto_tsquery(\'simple\', ?)', $builder->toSql());
-        $this->assertEquals(['Hello World'], $builder->getBindings());
-
-        $builder = $this->getPostgresBuilderWithProcessor();
-        $builder->select('*')->from('users')->whereFulltext(['body', 'title'], 'Car Plane');
-        $this->assertSame('select * from "users" where (to_tsvector(\'english\', "body") || to_tsvector(\'english\', "title")) @@ plainto_tsquery(\'english\', ?)', $builder->toSql());
-        $this->assertEquals(['Car Plane'], $builder->getBindings());
-    }
-
     public function testUnions()
     {
         $builder = $this->getBuilder();
         $builder->select('*')->from('users')->where('id', '=', 1);
         $builder->union($this->getBuilder()->select('*')->from('users')->where('id', '=', 2));
         $this->assertSame('(select * from "users" where "id" = ?) union (select * from "users" where "id" = ?)', $builder->toSql());
-        $this->assertEquals([0 => 1, 1 => 2], $builder->getBindings());
-
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->where('id', '=', 1);
-        $builder->union($this->getMySqlBuilder()->select('*')->from('users')->where('id', '=', 2));
-        $this->assertSame('(select * from `users` where `id` = ?) union (select * from `users` where `id` = ?)', $builder->toSql());
-        $this->assertEquals([0 => 1, 1 => 2], $builder->getBindings());
-
-        $builder = $this->getMysqlBuilder();
-        $expectedSql = '(select `a` from `t1` where `a` = ? and `b` = ?) union (select `a` from `t2` where `a` = ? and `b` = ?) order by `a` asc limit 10';
-        $union = $this->getMysqlBuilder()->select('a')->from('t2')->where('a', 11)->where('b', 2);
-        $builder->select('a')->from('t1')->where('a', 10)->where('b', 1)->union($union)->orderBy('a')->limit(10);
-        $this->assertEquals($expectedSql, $builder->toSql());
-        $this->assertEquals([0 => 10, 1 => 1, 2 => 11, 3 => 2], $builder->getBindings());
-
-        $builder = $this->getPostgresBuilder();
-        $expectedSql = '(select "name" from "users" where "id" = ?) union (select "name" from "users" where "id" = ?)';
-        $builder->select('name')->from('users')->where('id', '=', 1);
-        $builder->union($this->getPostgresBuilder()->select('name')->from('users')->where('id', '=', 2));
-        $this->assertEquals($expectedSql, $builder->toSql());
-        $this->assertEquals([0 => 1, 1 => 2], $builder->getBindings());
-
-        $builder = $this->getSQLiteBuilder();
-        $expectedSql = 'select * from (select "name" from "users" where "id" = ?) union select * from (select "name" from "users" where "id" = ?)';
-        $builder->select('name')->from('users')->where('id', '=', 1);
-        $builder->union($this->getSQLiteBuilder()->select('name')->from('users')->where('id', '=', 2));
-        $this->assertEquals($expectedSql, $builder->toSql());
-        $this->assertEquals([0 => 1, 1 => 2], $builder->getBindings());
-
-        $builder = $this->getSqlServerBuilder();
-        $expectedSql = 'select * from (select [name] from [users] where [id] = ?) as [temp_table] union select * from (select [name] from [users] where [id] = ?) as [temp_table]';
-        $builder->select('name')->from('users')->where('id', '=', 1);
-        $builder->union($this->getSqlServerBuilder()->select('name')->from('users')->where('id', '=', 2));
-        $this->assertEquals($expectedSql, $builder->toSql());
         $this->assertEquals([0 => 1, 1 => 2], $builder->getBindings());
     }
 
@@ -1001,13 +575,6 @@ class DatabaseQueryBuilderTest extends TestCase
         $builder->select('*')->from('users')->where('id', '=', 1);
         $builder->unionAll($this->getBuilder()->select('*')->from('users')->where('id', '=', 2));
         $this->assertSame('(select * from "users" where "id" = ?) union all (select * from "users" where "id" = ?)', $builder->toSql());
-        $this->assertEquals([0 => 1, 1 => 2], $builder->getBindings());
-
-        $expectedSql = '(select * from "users" where "id" = ?) union all (select * from "users" where "id" = ?)';
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->where('id', '=', 1);
-        $builder->unionAll($this->getBuilder()->select('*')->from('users')->where('id', '=', 2));
-        $this->assertEquals($expectedSql, $builder->toSql());
         $this->assertEquals([0 => 1, 1 => 2], $builder->getBindings());
     }
 
@@ -1048,20 +615,6 @@ class DatabaseQueryBuilderTest extends TestCase
         $builder->union($this->getBuilder()->select('*')->from('dogs'));
         $builder->skip(5)->take(10);
         $this->assertSame('(select * from "users") union (select * from "dogs") limit 10 offset 5', $builder->toSql());
-
-        $expectedSql = '(select * from "users") union (select * from "dogs") limit 10 offset 5';
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users');
-        $builder->union($this->getBuilder()->select('*')->from('dogs'));
-        $builder->skip(5)->take(10);
-        $this->assertEquals($expectedSql, $builder->toSql());
-
-        $expectedSql = '(select * from "users" limit 11) union (select * from "dogs" limit 22) limit 10 offset 5';
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->limit(11);
-        $builder->union($this->getBuilder()->select('*')->from('dogs')->limit(22));
-        $builder->skip(5)->take(10);
-        $this->assertEquals($expectedSql, $builder->toSql());
     }
 
     public function testUnionWithJoin()
@@ -1074,74 +627,6 @@ class DatabaseQueryBuilderTest extends TestCase
         }));
         $this->assertSame('(select * from "users") union (select * from "dogs" inner join "breeds" on "dogs"."breed_id" = "breeds"."id" and "breeds"."is_native" = ?)', $builder->toSql());
         $this->assertEquals([0 => 1], $builder->getBindings());
-    }
-
-    public function testMySqlUnionOrderBys()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->where('id', '=', 1);
-        $builder->union($this->getMySqlBuilder()->select('*')->from('users')->where('id', '=', 2));
-        $builder->orderBy('id', 'desc');
-        $this->assertSame('(select * from `users` where `id` = ?) union (select * from `users` where `id` = ?) order by `id` desc', $builder->toSql());
-        $this->assertEquals([0 => 1, 1 => 2], $builder->getBindings());
-    }
-
-    public function testMySqlUnionLimitsAndOffsets()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users');
-        $builder->union($this->getMySqlBuilder()->select('*')->from('dogs'));
-        $builder->skip(5)->take(10);
-        $this->assertSame('(select * from `users`) union (select * from `dogs`) limit 10 offset 5', $builder->toSql());
-    }
-
-    public function testUnionAggregate()
-    {
-        $expected = 'select count(*) as aggregate from ((select * from `posts`) union (select * from `videos`)) as `temp_table`';
-        $builder = $this->getMySqlBuilder();
-        $builder->getConnection()->shouldReceive('select')->once()->with($expected, [], true);
-        $builder->getProcessor()->shouldReceive('processSelect')->once();
-        $builder->from('posts')->union($this->getMySqlBuilder()->from('videos'))->count();
-
-        $expected = 'select count(*) as aggregate from ((select `id` from `posts`) union (select `id` from `videos`)) as `temp_table`';
-        $builder = $this->getMySqlBuilder();
-        $builder->getConnection()->shouldReceive('select')->once()->with($expected, [], true);
-        $builder->getProcessor()->shouldReceive('processSelect')->once();
-        $builder->from('posts')->select('id')->union($this->getMySqlBuilder()->from('videos')->select('id'))->count();
-
-        $expected = 'select count(*) as aggregate from ((select * from "posts") union (select * from "videos")) as "temp_table"';
-        $builder = $this->getPostgresBuilder();
-        $builder->getConnection()->shouldReceive('select')->once()->with($expected, [], true);
-        $builder->getProcessor()->shouldReceive('processSelect')->once();
-        $builder->from('posts')->union($this->getPostgresBuilder()->from('videos'))->count();
-
-        $expected = 'select count(*) as aggregate from (select * from (select * from "posts") union select * from (select * from "videos")) as "temp_table"';
-        $builder = $this->getSQLiteBuilder();
-        $builder->getConnection()->shouldReceive('select')->once()->with($expected, [], true);
-        $builder->getProcessor()->shouldReceive('processSelect')->once();
-        $builder->from('posts')->union($this->getSQLiteBuilder()->from('videos'))->count();
-
-        $expected = 'select count(*) as aggregate from (select * from (select * from [posts]) as [temp_table] union select * from (select * from [videos]) as [temp_table]) as [temp_table]';
-        $builder = $this->getSqlServerBuilder();
-        $builder->getConnection()->shouldReceive('select')->once()->with($expected, [], true);
-        $builder->getProcessor()->shouldReceive('processSelect')->once();
-        $builder->from('posts')->union($this->getSqlServerBuilder()->from('videos'))->count();
-    }
-
-    public function testHavingAggregate()
-    {
-        $expected = 'select count(*) as aggregate from (select (select `count(*)` from `videos` where `posts`.`id` = `videos`.`post_id`) as `videos_count` from `posts` having `videos_count` > ?) as `temp_table`';
-        $builder = $this->getMySqlBuilder();
-        $builder->getConnection()->shouldReceive('getDatabaseName');
-        $builder->getConnection()->shouldReceive('select')->once()->with($expected, [0 => 1], true)->andReturn([['aggregate' => 1]]);
-        $builder->getProcessor()->shouldReceive('processSelect')->once()->andReturnUsing(function ($builder, $results) {
-            return $results;
-        });
-
-        $builder->from('posts')->selectSub(function ($query) {
-            $query->from('videos')->select('count(*)')->whereColumn('posts.id', '=', 'videos.post_id');
-        }, 'videos_count')->having('videos_count', '>', 1);
-        $builder->count();
     }
 
     public function testSubSelectWhereIns()
@@ -1172,20 +657,6 @@ class DatabaseQueryBuilderTest extends TestCase
         $builder->select('*')->from('users')->where('id', '=', 1)->orWhereNull('id');
         $this->assertSame('select * from "users" where "id" = ? or "id" is null', $builder->toSql());
         $this->assertEquals([0 => 1], $builder->getBindings());
-    }
-
-    public function testJsonWhereNullMysql()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->whereNull('items->id');
-        $this->assertSame('select * from `users` where (json_extract(`items`, \'$."id"\') is null OR json_type(json_extract(`items`, \'$."id"\')) = \'NULL\')', $builder->toSql());
-    }
-
-    public function testJsonWhereNotNullMysql()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->whereNotNull('items->id');
-        $this->assertSame('select * from `users` where (json_extract(`items`, \'$."id"\') is not null AND json_type(json_extract(`items`, \'$."id"\')) != \'NULL\')', $builder->toSql());
     }
 
     public function testArrayWhereNulls()
@@ -1282,40 +753,6 @@ class DatabaseQueryBuilderTest extends TestCase
             ->orderByRaw('field(category, ?, ?) asc', ['news', 'opinion']);
         $this->assertSame('(select * from "posts" where "public" = ?) union all (select * from "videos" where "public" = ?) order by field(category, ?, ?) asc', $builder->toSql());
         $this->assertEquals([1, 1, 'news', 'opinion'], $builder->getBindings());
-    }
-
-    public function testOrderBysSqlServer()
-    {
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->orderBy('email')->orderBy('age', 'desc');
-        $this->assertSame('select * from [users] order by [email] asc, [age] desc', $builder->toSql());
-
-        $builder->orders = null;
-        $this->assertSame('select * from [users]', $builder->toSql());
-
-        $builder->orders = [];
-        $this->assertSame('select * from [users]', $builder->toSql());
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->orderBy('email');
-        $this->assertSame('select * from [users] order by [email] asc', $builder->toSql());
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->orderByDesc('name');
-        $this->assertSame('select * from [users] order by [name] desc', $builder->toSql());
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->orderByRaw('[age] asc');
-        $this->assertSame('select * from [users] order by [age] asc', $builder->toSql());
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->orderBy('email')->orderByRaw('[age] ? desc', ['foo']);
-        $this->assertSame('select * from [users] order by [email] asc, [age] ? desc', $builder->toSql());
-        $this->assertEquals(['foo'], $builder->getBindings());
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->skip(25)->take(10)->orderByRaw('[email] desc');
-        $this->assertSame('select * from [users] order by [email] desc offset 25 rows fetch next 10 rows only', $builder->toSql());
     }
 
     public function testReorder()
@@ -2246,14 +1683,6 @@ class DatabaseQueryBuilderTest extends TestCase
         $this->assertEquals(1, $results);
     }
 
-    public function testSqlServerExists()
-    {
-        $builder = $this->getSqlServerBuilder();
-        $builder->getConnection()->shouldReceive('select')->once()->with('select top 1 1 [exists] from [users]', [], true)->andReturn([['exists' => 1]]);
-        $results = $builder->from('users')->exists();
-        $this->assertTrue($results);
-    }
-
     public function testExistsOr()
     {
         $builder = $this->getBuilder();
@@ -2405,38 +1834,6 @@ class DatabaseQueryBuilderTest extends TestCase
         $builder->from('users')->insertOrIgnore(['email' => 'foo']);
     }
 
-    public function testMySqlInsertOrIgnoreMethod()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->getConnection()->shouldReceive('affectingStatement')->once()->with('insert ignore into `users` (`email`) values (?)', ['foo'])->andReturn(1);
-        $result = $builder->from('users')->insertOrIgnore(['email' => 'foo']);
-        $this->assertEquals(1, $result);
-    }
-
-    public function testPostgresInsertOrIgnoreMethod()
-    {
-        $builder = $this->getPostgresBuilder();
-        $builder->getConnection()->shouldReceive('affectingStatement')->once()->with('insert into "users" ("email") values (?) on conflict do nothing', ['foo'])->andReturn(1);
-        $result = $builder->from('users')->insertOrIgnore(['email' => 'foo']);
-        $this->assertEquals(1, $result);
-    }
-
-    public function testSQLiteInsertOrIgnoreMethod()
-    {
-        $builder = $this->getSQLiteBuilder();
-        $builder->getConnection()->shouldReceive('affectingStatement')->once()->with('insert or ignore into "users" ("email") values (?)', ['foo'])->andReturn(1);
-        $result = $builder->from('users')->insertOrIgnore(['email' => 'foo']);
-        $this->assertEquals(1, $result);
-    }
-
-    public function testSqlServerInsertOrIgnoreMethod()
-    {
-        $this->expectException(RuntimeException::class);
-        $this->expectExceptionMessage('does not support');
-        $builder = $this->getSqlServerBuilder();
-        $builder->from('users')->insertOrIgnore(['email' => 'foo']);
-    }
-
     public function testInsertGetIdMethod()
     {
         $builder = $this->getBuilder();
@@ -2451,25 +1848,6 @@ class DatabaseQueryBuilderTest extends TestCase
         $builder->getProcessor()->shouldReceive('processInsertGetId')->once()->with($builder, 'insert into "users" ("email", "bar") values (?, bar)', ['foo'], 'id')->andReturn(1);
         $result = $builder->from('users')->insertGetId(['email' => 'foo', 'bar' => new Raw('bar')], 'id');
         $this->assertEquals(1, $result);
-    }
-
-    public function testInsertGetIdWithEmptyValues()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->getProcessor()->shouldReceive('processInsertGetId')->once()->with($builder, 'insert into `users` () values ()', [], null);
-        $builder->from('users')->insertGetId([]);
-
-        $builder = $this->getPostgresBuilder();
-        $builder->getProcessor()->shouldReceive('processInsertGetId')->once()->with($builder, 'insert into "users" default values returning "id"', [], null);
-        $builder->from('users')->insertGetId([]);
-
-        $builder = $this->getSQLiteBuilder();
-        $builder->getProcessor()->shouldReceive('processInsertGetId')->once()->with($builder, 'insert into "users" default values', [], null);
-        $builder->from('users')->insertGetId([]);
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->getProcessor()->shouldReceive('processInsertGetId')->once()->with($builder, 'insert into [users] default values', [], null);
-        $builder->from('users')->insertGetId([]);
     }
 
     public function testInsertMethodRespectsRawBindings()
@@ -2494,75 +1872,6 @@ class DatabaseQueryBuilderTest extends TestCase
         $builder->getConnection()->shouldReceive('update')->once()->with('update "users" set "email" = ?, "name" = ? where "id" = ?', ['foo', 'bar', 1])->andReturn(1);
         $result = $builder->from('users')->where('id', '=', 1)->update(['email' => 'foo', 'name' => 'bar']);
         $this->assertEquals(1, $result);
-
-        $builder = $this->getMySqlBuilder();
-        $builder->getConnection()->shouldReceive('update')->once()->with('update `users` set `email` = ?, `name` = ? where `id` = ? order by `foo` desc limit 5', ['foo', 'bar', 1])->andReturn(1);
-        $result = $builder->from('users')->where('id', '=', 1)->orderBy('foo', 'desc')->limit(5)->update(['email' => 'foo', 'name' => 'bar']);
-        $this->assertEquals(1, $result);
-    }
-
-    public function testUpsertMethod()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->getConnection()
-            ->shouldReceive('getConfig')->with('use_upsert_alias')->andReturn(false)
-            ->shouldReceive('affectingStatement')->once()->with('insert into `users` (`email`, `name`) values (?, ?), (?, ?) on duplicate key update `email` = values(`email`), `name` = values(`name`)', ['foo', 'bar', 'foo2', 'bar2'])->andReturn(2);
-        $result = $builder->from('users')->upsert([['email' => 'foo', 'name' => 'bar'], ['name' => 'bar2', 'email' => 'foo2']], 'email');
-        $this->assertEquals(2, $result);
-
-        $builder = $this->getMySqlBuilder();
-        $builder->getConnection()
-            ->shouldReceive('getConfig')->with('use_upsert_alias')->andReturn(true)
-            ->shouldReceive('affectingStatement')->once()->with('insert into `users` (`email`, `name`) values (?, ?), (?, ?) as laravel_upsert_alias on duplicate key update `email` = `laravel_upsert_alias`.`email`, `name` = `laravel_upsert_alias`.`name`', ['foo', 'bar', 'foo2', 'bar2'])->andReturn(2);
-        $result = $builder->from('users')->upsert([['email' => 'foo', 'name' => 'bar'], ['name' => 'bar2', 'email' => 'foo2']], 'email');
-        $this->assertEquals(2, $result);
-
-        $builder = $this->getPostgresBuilder();
-        $builder->getConnection()->shouldReceive('affectingStatement')->once()->with('insert into "users" ("email", "name") values (?, ?), (?, ?) on conflict ("email") do update set "email" = "excluded"."email", "name" = "excluded"."name"', ['foo', 'bar', 'foo2', 'bar2'])->andReturn(2);
-        $result = $builder->from('users')->upsert([['email' => 'foo', 'name' => 'bar'], ['name' => 'bar2', 'email' => 'foo2']], 'email');
-        $this->assertEquals(2, $result);
-
-        $builder = $this->getSQLiteBuilder();
-        $builder->getConnection()->shouldReceive('affectingStatement')->once()->with('insert into "users" ("email", "name") values (?, ?), (?, ?) on conflict ("email") do update set "email" = "excluded"."email", "name" = "excluded"."name"', ['foo', 'bar', 'foo2', 'bar2'])->andReturn(2);
-        $result = $builder->from('users')->upsert([['email' => 'foo', 'name' => 'bar'], ['name' => 'bar2', 'email' => 'foo2']], 'email');
-        $this->assertEquals(2, $result);
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->getConnection()->shouldReceive('affectingStatement')->once()->with('merge [users] using (values (?, ?), (?, ?)) [laravel_source] ([email], [name]) on [laravel_source].[email] = [users].[email] when matched then update set [email] = [laravel_source].[email], [name] = [laravel_source].[name] when not matched then insert ([email], [name]) values ([email], [name]);', ['foo', 'bar', 'foo2', 'bar2'])->andReturn(2);
-        $result = $builder->from('users')->upsert([['email' => 'foo', 'name' => 'bar'], ['name' => 'bar2', 'email' => 'foo2']], 'email');
-        $this->assertEquals(2, $result);
-    }
-
-    public function testUpsertMethodWithUpdateColumns()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->getConnection()
-            ->shouldReceive('getConfig')->with('use_upsert_alias')->andReturn(false)
-            ->shouldReceive('affectingStatement')->once()->with('insert into `users` (`email`, `name`) values (?, ?), (?, ?) on duplicate key update `name` = values(`name`)', ['foo', 'bar', 'foo2', 'bar2'])->andReturn(2);
-        $result = $builder->from('users')->upsert([['email' => 'foo', 'name' => 'bar'], ['name' => 'bar2', 'email' => 'foo2']], 'email', ['name']);
-        $this->assertEquals(2, $result);
-
-        $builder = $this->getMySqlBuilder();
-        $builder->getConnection()
-            ->shouldReceive('getConfig')->with('use_upsert_alias')->andReturn(true)
-            ->shouldReceive('affectingStatement')->once()->with('insert into `users` (`email`, `name`) values (?, ?), (?, ?) as laravel_upsert_alias on duplicate key update `name` = `laravel_upsert_alias`.`name`', ['foo', 'bar', 'foo2', 'bar2'])->andReturn(2);
-        $result = $builder->from('users')->upsert([['email' => 'foo', 'name' => 'bar'], ['name' => 'bar2', 'email' => 'foo2']], 'email', ['name']);
-        $this->assertEquals(2, $result);
-
-        $builder = $this->getPostgresBuilder();
-        $builder->getConnection()->shouldReceive('affectingStatement')->once()->with('insert into "users" ("email", "name") values (?, ?), (?, ?) on conflict ("email") do update set "name" = "excluded"."name"', ['foo', 'bar', 'foo2', 'bar2'])->andReturn(2);
-        $result = $builder->from('users')->upsert([['email' => 'foo', 'name' => 'bar'], ['name' => 'bar2', 'email' => 'foo2']], 'email', ['name']);
-        $this->assertEquals(2, $result);
-
-        $builder = $this->getSQLiteBuilder();
-        $builder->getConnection()->shouldReceive('affectingStatement')->once()->with('insert into "users" ("email", "name") values (?, ?), (?, ?) on conflict ("email") do update set "name" = "excluded"."name"', ['foo', 'bar', 'foo2', 'bar2'])->andReturn(2);
-        $result = $builder->from('users')->upsert([['email' => 'foo', 'name' => 'bar'], ['name' => 'bar2', 'email' => 'foo2']], 'email', ['name']);
-        $this->assertEquals(2, $result);
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->getConnection()->shouldReceive('affectingStatement')->once()->with('merge [users] using (values (?, ?), (?, ?)) [laravel_source] ([email], [name]) on [laravel_source].[email] = [users].[email] when matched then update set [name] = [laravel_source].[name] when not matched then insert ([email], [name]) values ([email], [name]);', ['foo', 'bar', 'foo2', 'bar2'])->andReturn(2);
-        $result = $builder->from('users')->upsert([['email' => 'foo', 'name' => 'bar'], ['name' => 'bar2', 'email' => 'foo2']], 'email', ['name']);
-        $this->assertEquals(2, $result);
     }
 
     public function testUpdateMethodWithJoins()
@@ -2578,142 +1887,6 @@ class DatabaseQueryBuilderTest extends TestCase
             $join->on('users.id', '=', 'orders.user_id')
                 ->where('users.id', '=', 1);
         })->update(['email' => 'foo', 'name' => 'bar']);
-        $this->assertEquals(1, $result);
-    }
-
-    public function testUpdateMethodWithJoinsOnSqlServer()
-    {
-        $builder = $this->getSqlServerBuilder();
-        $builder->getConnection()->shouldReceive('update')->once()->with('update [users] set [email] = ?, [name] = ? from [users] inner join [orders] on [users].[id] = [orders].[user_id] where [users].[id] = ?', ['foo', 'bar', 1])->andReturn(1);
-        $result = $builder->from('users')->join('orders', 'users.id', '=', 'orders.user_id')->where('users.id', '=', 1)->update(['email' => 'foo', 'name' => 'bar']);
-        $this->assertEquals(1, $result);
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->getConnection()->shouldReceive('update')->once()->with('update [users] set [email] = ?, [name] = ? from [users] inner join [orders] on [users].[id] = [orders].[user_id] and [users].[id] = ?', ['foo', 'bar', 1])->andReturn(1);
-        $result = $builder->from('users')->join('orders', function ($join) {
-            $join->on('users.id', '=', 'orders.user_id')
-                ->where('users.id', '=', 1);
-        })->update(['email' => 'foo', 'name' => 'bar']);
-        $this->assertEquals(1, $result);
-    }
-
-    public function testUpdateMethodWithJoinsOnMySql()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->getConnection()->shouldReceive('update')->once()->with('update `users` inner join `orders` on `users`.`id` = `orders`.`user_id` set `email` = ?, `name` = ? where `users`.`id` = ?', ['foo', 'bar', 1])->andReturn(1);
-        $result = $builder->from('users')->join('orders', 'users.id', '=', 'orders.user_id')->where('users.id', '=', 1)->update(['email' => 'foo', 'name' => 'bar']);
-        $this->assertEquals(1, $result);
-
-        $builder = $this->getMySqlBuilder();
-        $builder->getConnection()->shouldReceive('update')->once()->with('update `users` inner join `orders` on `users`.`id` = `orders`.`user_id` and `users`.`id` = ? set `email` = ?, `name` = ?', [1, 'foo', 'bar'])->andReturn(1);
-        $result = $builder->from('users')->join('orders', function ($join) {
-            $join->on('users.id', '=', 'orders.user_id')
-                ->where('users.id', '=', 1);
-        })->update(['email' => 'foo', 'name' => 'bar']);
-        $this->assertEquals(1, $result);
-    }
-
-    public function testUpdateMethodWithJoinsOnSQLite()
-    {
-        $builder = $this->getSQLiteBuilder();
-        $builder->getConnection()->shouldReceive('update')->once()->with('update "users" set "email" = ?, "name" = ? where "rowid" in (select "users"."rowid" from "users" where "users"."id" > ? order by "id" asc limit 3)', ['foo', 'bar', 1])->andReturn(1);
-        $result = $builder->from('users')->where('users.id', '>', 1)->limit(3)->oldest('id')->update(['email' => 'foo', 'name' => 'bar']);
-        $this->assertEquals(1, $result);
-
-        $builder = $this->getSQLiteBuilder();
-        $builder->getConnection()->shouldReceive('update')->once()->with('update "users" set "email" = ?, "name" = ? where "rowid" in (select "users"."rowid" from "users" inner join "orders" on "users"."id" = "orders"."user_id" where "users"."id" = ?)', ['foo', 'bar', 1])->andReturn(1);
-        $result = $builder->from('users')->join('orders', 'users.id', '=', 'orders.user_id')->where('users.id', '=', 1)->update(['email' => 'foo', 'name' => 'bar']);
-        $this->assertEquals(1, $result);
-
-        $builder = $this->getSQLiteBuilder();
-        $builder->getConnection()->shouldReceive('update')->once()->with('update "users" set "email" = ?, "name" = ? where "rowid" in (select "users"."rowid" from "users" inner join "orders" on "users"."id" = "orders"."user_id" and "users"."id" = ?)', ['foo', 'bar', 1])->andReturn(1);
-        $result = $builder->from('users')->join('orders', function ($join) {
-            $join->on('users.id', '=', 'orders.user_id')
-                ->where('users.id', '=', 1);
-        })->update(['email' => 'foo', 'name' => 'bar']);
-        $this->assertEquals(1, $result);
-
-        $builder = $this->getSQLiteBuilder();
-        $builder->getConnection()->shouldReceive('update')->once()->with('update "users" as "u" set "email" = ?, "name" = ? where "rowid" in (select "u"."rowid" from "users" as "u" inner join "orders" as "o" on "u"."id" = "o"."user_id")', ['foo', 'bar'])->andReturn(1);
-        $result = $builder->from('users as u')->join('orders as o', 'u.id', '=', 'o.user_id')->update(['email' => 'foo', 'name' => 'bar']);
-        $this->assertEquals(1, $result);
-    }
-
-    public function testUpdateMethodWithJoinsAndAliasesOnSqlServer()
-    {
-        $builder = $this->getSqlServerBuilder();
-        $builder->getConnection()->shouldReceive('update')->once()->with('update [u] set [email] = ?, [name] = ? from [users] as [u] inner join [orders] on [u].[id] = [orders].[user_id] where [u].[id] = ?', ['foo', 'bar', 1])->andReturn(1);
-        $result = $builder->from('users as u')->join('orders', 'u.id', '=', 'orders.user_id')->where('u.id', '=', 1)->update(['email' => 'foo', 'name' => 'bar']);
-        $this->assertEquals(1, $result);
-    }
-
-    public function testUpdateMethodWithoutJoinsOnPostgres()
-    {
-        $builder = $this->getPostgresBuilder();
-        $builder->getConnection()->shouldReceive('update')->once()->with('update "users" set "email" = ?, "name" = ? where "id" = ?', ['foo', 'bar', 1])->andReturn(1);
-        $result = $builder->from('users')->where('id', '=', 1)->update(['users.email' => 'foo', 'name' => 'bar']);
-        $this->assertEquals(1, $result);
-
-        $builder = $this->getPostgresBuilder();
-        $builder->getConnection()->shouldReceive('update')->once()->with('update "users" set "email" = ?, "name" = ? where "id" = ?', ['foo', 'bar', 1])->andReturn(1);
-        $result = $builder->from('users')->where('id', '=', 1)->selectRaw('?', ['ignore'])->update(['users.email' => 'foo', 'name' => 'bar']);
-        $this->assertEquals(1, $result);
-
-        $builder = $this->getPostgresBuilder();
-        $builder->getConnection()->shouldReceive('update')->once()->with('update "users"."users" set "email" = ?, "name" = ? where "id" = ?', ['foo', 'bar', 1])->andReturn(1);
-        $result = $builder->from('users.users')->where('id', '=', 1)->selectRaw('?', ['ignore'])->update(['users.users.email' => 'foo', 'name' => 'bar']);
-        $this->assertEquals(1, $result);
-    }
-
-    public function testUpdateMethodWithJoinsOnPostgres()
-    {
-        $builder = $this->getPostgresBuilder();
-        $builder->getConnection()->shouldReceive('update')->once()->with('update "users" set "email" = ?, "name" = ? where "ctid" in (select "users"."ctid" from "users" inner join "orders" on "users"."id" = "orders"."user_id" where "users"."id" = ?)', ['foo', 'bar', 1])->andReturn(1);
-        $result = $builder->from('users')->join('orders', 'users.id', '=', 'orders.user_id')->where('users.id', '=', 1)->update(['email' => 'foo', 'name' => 'bar']);
-        $this->assertEquals(1, $result);
-
-        $builder = $this->getPostgresBuilder();
-        $builder->getConnection()->shouldReceive('update')->once()->with('update "users" set "email" = ?, "name" = ? where "ctid" in (select "users"."ctid" from "users" inner join "orders" on "users"."id" = "orders"."user_id" and "users"."id" = ?)', ['foo', 'bar', 1])->andReturn(1);
-        $result = $builder->from('users')->join('orders', function ($join) {
-            $join->on('users.id', '=', 'orders.user_id')
-                ->where('users.id', '=', 1);
-        })->update(['email' => 'foo', 'name' => 'bar']);
-        $this->assertEquals(1, $result);
-
-        $builder = $this->getPostgresBuilder();
-        $builder->getConnection()->shouldReceive('update')->once()->with('update "users" set "email" = ?, "name" = ? where "ctid" in (select "users"."ctid" from "users" inner join "orders" on "users"."id" = "orders"."user_id" and "users"."id" = ? where "name" = ?)', ['foo', 'bar', 1, 'baz'])->andReturn(1);
-        $result = $builder->from('users')
-            ->join('orders', function ($join) {
-                $join->on('users.id', '=', 'orders.user_id')
-                    ->where('users.id', '=', 1);
-            })->where('name', 'baz')
-            ->update(['email' => 'foo', 'name' => 'bar']);
-        $this->assertEquals(1, $result);
-    }
-
-    public function testUpdateFromMethodWithJoinsOnPostgres()
-    {
-        $builder = $this->getPostgresBuilder();
-        $builder->getConnection()->shouldReceive('update')->once()->with('update "users" set "email" = ?, "name" = ? from "orders" where "users"."id" = ? and "users"."id" = "orders"."user_id"', ['foo', 'bar', 1])->andReturn(1);
-        $result = $builder->from('users')->join('orders', 'users.id', '=', 'orders.user_id')->where('users.id', '=', 1)->updateFrom(['email' => 'foo', 'name' => 'bar']);
-        $this->assertEquals(1, $result);
-
-        $builder = $this->getPostgresBuilder();
-        $builder->getConnection()->shouldReceive('update')->once()->with('update "users" set "email" = ?, "name" = ? from "orders" where "users"."id" = "orders"."user_id" and "users"."id" = ?', ['foo', 'bar', 1])->andReturn(1);
-        $result = $builder->from('users')->join('orders', function ($join) {
-            $join->on('users.id', '=', 'orders.user_id')
-                ->where('users.id', '=', 1);
-        })->updateFrom(['email' => 'foo', 'name' => 'bar']);
-        $this->assertEquals(1, $result);
-
-        $builder = $this->getPostgresBuilder();
-        $builder->getConnection()->shouldReceive('update')->once()->with('update "users" set "email" = ?, "name" = ? from "orders" where "name" = ? and "users"."id" = "orders"."user_id" and "users"."id" = ?', ['foo', 'bar', 'baz', 1])->andReturn(1);
-        $result = $builder->from('users')
-            ->join('orders', function ($join) {
-                $join->on('users.id', '=', 'orders.user_id')
-                   ->where('users.id', '=', 1);
-            })->where('name', 'baz')
-           ->updateFrom(['email' => 'foo', 'name' => 'bar']);
         $this->assertEquals(1, $result);
     }
 
@@ -2784,99 +1957,6 @@ class DatabaseQueryBuilderTest extends TestCase
         $builder->getConnection()->shouldReceive('delete')->once()->with('delete from "users" where "users"."id" = ?', [1])->andReturn(1);
         $result = $builder->from('users')->selectRaw('?', ['ignore'])->delete(1);
         $this->assertEquals(1, $result);
-
-        $builder = $this->getSqliteBuilder();
-        $builder->getConnection()->shouldReceive('delete')->once()->with('delete from "users" where "rowid" in (select "users"."rowid" from "users" where "email" = ? order by "id" asc limit 1)', ['foo'])->andReturn(1);
-        $result = $builder->from('users')->where('email', '=', 'foo')->orderBy('id')->take(1)->delete();
-        $this->assertEquals(1, $result);
-
-        $builder = $this->getMySqlBuilder();
-        $builder->getConnection()->shouldReceive('delete')->once()->with('delete from `users` where `email` = ? order by `id` asc limit 1', ['foo'])->andReturn(1);
-        $result = $builder->from('users')->where('email', '=', 'foo')->orderBy('id')->take(1)->delete();
-        $this->assertEquals(1, $result);
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->getConnection()->shouldReceive('delete')->once()->with('delete from [users] where [email] = ?', ['foo'])->andReturn(1);
-        $result = $builder->from('users')->where('email', '=', 'foo')->delete();
-        $this->assertEquals(1, $result);
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->getConnection()->shouldReceive('delete')->once()->with('delete top (1) from [users] where [email] = ?', ['foo'])->andReturn(1);
-        $result = $builder->from('users')->where('email', '=', 'foo')->orderBy('id')->take(1)->delete();
-        $this->assertEquals(1, $result);
-    }
-
-    public function testDeleteWithJoinMethod()
-    {
-        $builder = $this->getSqliteBuilder();
-        $builder->getConnection()->shouldReceive('delete')->once()->with('delete from "users" where "rowid" in (select "users"."rowid" from "users" inner join "contacts" on "users"."id" = "contacts"."id" where "users"."email" = ? order by "users"."id" asc limit 1)', ['foo'])->andReturn(1);
-        $result = $builder->from('users')->join('contacts', 'users.id', '=', 'contacts.id')->where('users.email', '=', 'foo')->orderBy('users.id')->limit(1)->delete();
-        $this->assertEquals(1, $result);
-
-        $builder = $this->getSqliteBuilder();
-        $builder->getConnection()->shouldReceive('delete')->once()->with('delete from "users" as "u" where "rowid" in (select "u"."rowid" from "users" as "u" inner join "contacts" as "c" on "u"."id" = "c"."id")', [])->andReturn(1);
-        $result = $builder->from('users as u')->join('contacts as c', 'u.id', '=', 'c.id')->delete();
-        $this->assertEquals(1, $result);
-
-        $builder = $this->getMySqlBuilder();
-        $builder->getConnection()->shouldReceive('delete')->once()->with('delete `users` from `users` inner join `contacts` on `users`.`id` = `contacts`.`id` where `email` = ?', ['foo'])->andReturn(1);
-        $result = $builder->from('users')->join('contacts', 'users.id', '=', 'contacts.id')->where('email', '=', 'foo')->orderBy('id')->limit(1)->delete();
-        $this->assertEquals(1, $result);
-
-        $builder = $this->getMySqlBuilder();
-        $builder->getConnection()->shouldReceive('delete')->once()->with('delete `a` from `users` as `a` inner join `users` as `b` on `a`.`id` = `b`.`user_id` where `email` = ?', ['foo'])->andReturn(1);
-        $result = $builder->from('users AS a')->join('users AS b', 'a.id', '=', 'b.user_id')->where('email', '=', 'foo')->orderBy('id')->limit(1)->delete();
-        $this->assertEquals(1, $result);
-
-        $builder = $this->getMySqlBuilder();
-        $builder->getConnection()->shouldReceive('delete')->once()->with('delete `users` from `users` inner join `contacts` on `users`.`id` = `contacts`.`id` where `users`.`id` = ?', [1])->andReturn(1);
-        $result = $builder->from('users')->join('contacts', 'users.id', '=', 'contacts.id')->orderBy('id')->take(1)->delete(1);
-        $this->assertEquals(1, $result);
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->getConnection()->shouldReceive('delete')->once()->with('delete [users] from [users] inner join [contacts] on [users].[id] = [contacts].[id] where [email] = ?', ['foo'])->andReturn(1);
-        $result = $builder->from('users')->join('contacts', 'users.id', '=', 'contacts.id')->where('email', '=', 'foo')->delete();
-        $this->assertEquals(1, $result);
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->getConnection()->shouldReceive('delete')->once()->with('delete [a] from [users] as [a] inner join [users] as [b] on [a].[id] = [b].[user_id] where [email] = ?', ['foo'])->andReturn(1);
-        $result = $builder->from('users AS a')->join('users AS b', 'a.id', '=', 'b.user_id')->where('email', '=', 'foo')->orderBy('id')->limit(1)->delete();
-        $this->assertEquals(1, $result);
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->getConnection()->shouldReceive('delete')->once()->with('delete [users] from [users] inner join [contacts] on [users].[id] = [contacts].[id] where [users].[id] = ?', [1])->andReturn(1);
-        $result = $builder->from('users')->join('contacts', 'users.id', '=', 'contacts.id')->delete(1);
-        $this->assertEquals(1, $result);
-
-        $builder = $this->getPostgresBuilder();
-        $builder->getConnection()->shouldReceive('delete')->once()->with('delete from "users" where "ctid" in (select "users"."ctid" from "users" inner join "contacts" on "users"."id" = "contacts"."id" where "users"."email" = ?)', ['foo'])->andReturn(1);
-        $result = $builder->from('users')->join('contacts', 'users.id', '=', 'contacts.id')->where('users.email', '=', 'foo')->delete();
-        $this->assertEquals(1, $result);
-
-        $builder = $this->getPostgresBuilder();
-        $builder->getConnection()->shouldReceive('delete')->once()->with('delete from "users" as "a" where "ctid" in (select "a"."ctid" from "users" as "a" inner join "users" as "b" on "a"."id" = "b"."user_id" where "email" = ? order by "id" asc limit 1)', ['foo'])->andReturn(1);
-        $result = $builder->from('users AS a')->join('users AS b', 'a.id', '=', 'b.user_id')->where('email', '=', 'foo')->orderBy('id')->limit(1)->delete();
-        $this->assertEquals(1, $result);
-
-        $builder = $this->getPostgresBuilder();
-        $builder->getConnection()->shouldReceive('delete')->once()->with('delete from "users" where "ctid" in (select "users"."ctid" from "users" inner join "contacts" on "users"."id" = "contacts"."id" where "users"."id" = ? order by "id" asc limit 1)', [1])->andReturn(1);
-        $result = $builder->from('users')->join('contacts', 'users.id', '=', 'contacts.id')->orderBy('id')->take(1)->delete(1);
-        $this->assertEquals(1, $result);
-
-        $builder = $this->getPostgresBuilder();
-        $builder->getConnection()->shouldReceive('delete')->once()->with('delete from "users" where "ctid" in (select "users"."ctid" from "users" inner join "contacts" on "users"."id" = "contacts"."user_id" and "users"."id" = ? where "name" = ?)', [1, 'baz'])->andReturn(1);
-        $result = $builder->from('users')
-            ->join('contacts', function ($join) {
-                $join->on('users.id', '=', 'contacts.user_id')
-                    ->where('users.id', '=', 1);
-            })->where('name', 'baz')
-            ->delete();
-        $this->assertEquals(1, $result);
-
-        $builder = $this->getPostgresBuilder();
-        $builder->getConnection()->shouldReceive('delete')->once()->with('delete from "users" where "ctid" in (select "users"."ctid" from "users" inner join "contacts" on "users"."id" = "contacts"."id")', [])->andReturn(1);
-        $result = $builder->from('users')->join('contacts', 'users.id', '=', 'contacts.id')->delete();
-        $this->assertEquals(1, $result);
     }
 
     public function testTruncateMethod()
@@ -2884,14 +1964,6 @@ class DatabaseQueryBuilderTest extends TestCase
         $builder = $this->getBuilder();
         $builder->getConnection()->shouldReceive('statement')->once()->with('truncate table "users"', []);
         $builder->from('users')->truncate();
-
-        $sqlite = new SQLiteGrammar;
-        $builder = $this->getBuilder();
-        $builder->from('users');
-        $this->assertEquals([
-            'delete from sqlite_sequence where name = ?' => ['users'],
-            'delete from "users"' => [],
-        ], $sqlite->compileTruncate($builder));
     }
 
     public function testPreserveAddsClosureToArray()
@@ -2954,27 +2026,6 @@ class DatabaseQueryBuilderTest extends TestCase
         $builder->insertUsing([], $this->getBuilder());
     }
 
-    public function testPreservedAreAppliedByUpsert()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->getConnection()
-            ->shouldReceive('getConfig')->with('use_upsert_alias')->andReturn(false)
-            ->shouldReceive('affectingStatement')->once()->with('insert into `users` (`email`) values (?) on duplicate key update `email` = values(`email`)', ['foo']);
-        $builder->beforeQuery(function ($builder) {
-            $builder->from('users');
-        });
-        $builder->upsert(['email' => 'foo'], 'id');
-
-        $builder = $this->getMySqlBuilder();
-        $builder->getConnection()
-            ->shouldReceive('getConfig')->with('use_upsert_alias')->andReturn(true)
-            ->shouldReceive('affectingStatement')->once()->with('insert into `users` (`email`) values (?) as laravel_upsert_alias on duplicate key update `email` = `laravel_upsert_alias`.`email`', ['foo']);
-        $builder->beforeQuery(function ($builder) {
-            $builder->from('users');
-        });
-        $builder->upsert(['email' => 'foo'], 'id');
-    }
-
     public function testPreservedAreAppliedByUpdate()
     {
         $builder = $this->getBuilder();
@@ -3015,468 +2066,15 @@ class DatabaseQueryBuilderTest extends TestCase
         $builder->exists();
     }
 
-    public function testPostgresInsertGetId()
-    {
-        $builder = $this->getPostgresBuilder();
-        $builder->getProcessor()->shouldReceive('processInsertGetId')->once()->with($builder, 'insert into "users" ("email") values (?) returning "id"', ['foo'], 'id')->andReturn(1);
-        $result = $builder->from('users')->insertGetId(['email' => 'foo'], 'id');
-        $this->assertEquals(1, $result);
-    }
-
-    public function testMySqlWrapping()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users');
-        $this->assertSame('select * from `users`', $builder->toSql());
-    }
-
-    public function testMySqlUpdateWrappingJson()
-    {
-        $grammar = new MySqlGrammar;
-        $processor = m::mock(Processor::class);
-
-        $connection = $this->createMock(ConnectionInterface::class);
-        $connection->expects($this->once())
-                    ->method('update')
-                    ->with(
-                        'update `users` set `name` = json_set(`name`, \'$."first_name"\', ?), `name` = json_set(`name`, \'$."last_name"\', ?) where `active` = ?',
-                        ['John', 'Doe', 1]
-                    );
-
-        $builder = new Builder($connection, $grammar, $processor);
-
-        $builder->from('users')->where('active', '=', 1)->update(['name->first_name' => 'John', 'name->last_name' => 'Doe']);
-    }
-
-    public function testMySqlUpdateWrappingNestedJson()
-    {
-        $grammar = new MySqlGrammar;
-        $processor = m::mock(Processor::class);
-
-        $connection = $this->createMock(ConnectionInterface::class);
-        $connection->expects($this->once())
-                    ->method('update')
-                    ->with(
-                        'update `users` set `meta` = json_set(`meta`, \'$."name"."first_name"\', ?), `meta` = json_set(`meta`, \'$."name"."last_name"\', ?) where `active` = ?',
-                        ['John', 'Doe', 1]
-                    );
-
-        $builder = new Builder($connection, $grammar, $processor);
-
-        $builder->from('users')->where('active', '=', 1)->update(['meta->name->first_name' => 'John', 'meta->name->last_name' => 'Doe']);
-    }
-
-    public function testMySqlUpdateWrappingJsonArray()
-    {
-        $grammar = new MySqlGrammar;
-        $processor = m::mock(Processor::class);
-
-        $connection = $this->createMock(ConnectionInterface::class);
-        $connection->expects($this->once())
-                    ->method('update')
-                    ->with(
-                        'update `users` set `options` = ?, `meta` = json_set(`meta`, \'$."tags"\', cast(? as json)), `group_id` = 45, `created_at` = ? where `active` = ?',
-                        [
-                            json_encode(['2fa' => false, 'presets' => ['laravel', 'vue']]),
-                            json_encode(['white', 'large']),
-                            new DateTime('2019-08-06'),
-                            1,
-                        ]
-                    );
-
-        $builder = new Builder($connection, $grammar, $processor);
-        $builder->from('users')->where('active', 1)->update([
-            'options' => ['2fa' => false, 'presets' => ['laravel', 'vue']],
-            'meta->tags' => ['white', 'large'],
-            'group_id' => new Raw('45'),
-            'created_at' => new DateTime('2019-08-06'),
-        ]);
-    }
-
-    public function testMySqlUpdateWrappingJsonPathArrayIndex()
-    {
-        $grammar = new MySqlGrammar;
-        $processor = m::mock(Processor::class);
-
-        $connection = $this->createMock(ConnectionInterface::class);
-        $connection->expects($this->once())
-                    ->method('update')
-                    ->with(
-                        'update `users` set `options` = json_set(`options`, \'$[1]."2fa"\', false), `meta` = json_set(`meta`, \'$."tags"[0][2]\', ?) where `active` = ?',
-                        [
-                            'large',
-                            1,
-                        ]
-                    );
-
-        $builder = new Builder($connection, $grammar, $processor);
-        $builder->from('users')->where('active', 1)->update([
-            'options->[1]->2fa' => false,
-            'meta->tags[0][2]' => 'large',
-        ]);
-    }
-
-    public function testMySqlUpdateWithJsonPreparesBindingsCorrectly()
-    {
-        $grammar = new MySqlGrammar;
-        $processor = m::mock(Processor::class);
-
-        $connection = m::mock(ConnectionInterface::class);
-        $connection->shouldReceive('update')
-                    ->once()
-                    ->with(
-                        'update `users` set `options` = json_set(`options`, \'$."enable"\', false), `updated_at` = ? where `id` = ?',
-                        ['2015-05-26 22:02:06', 0]
-                    );
-        $builder = new Builder($connection, $grammar, $processor);
-        $builder->from('users')->where('id', '=', 0)->update(['options->enable' => false, 'updated_at' => '2015-05-26 22:02:06']);
-
-        $connection->shouldReceive('update')
-            ->once()
-            ->with(
-                'update `users` set `options` = json_set(`options`, \'$."size"\', ?), `updated_at` = ? where `id` = ?',
-                [45, '2015-05-26 22:02:06', 0]
-            );
-        $builder = new Builder($connection, $grammar, $processor);
-        $builder->from('users')->where('id', '=', 0)->update(['options->size' => 45, 'updated_at' => '2015-05-26 22:02:06']);
-
-        $builder = $this->getMySqlBuilder();
-        $builder->getConnection()->shouldReceive('update')->once()->with('update `users` set `options` = json_set(`options`, \'$."size"\', ?)', [null]);
-        $builder->from('users')->update(['options->size' => null]);
-
-        $builder = $this->getMySqlBuilder();
-        $builder->getConnection()->shouldReceive('update')->once()->with('update `users` set `options` = json_set(`options`, \'$."size"\', 45)', []);
-        $builder->from('users')->update(['options->size' => new Raw('45')]);
-    }
-
-    public function testPostgresUpdateWrappingJson()
-    {
-        $builder = $this->getPostgresBuilder();
-        $builder->getConnection()->shouldReceive('update')
-            ->with('update "users" set "options" = jsonb_set("options"::jsonb, \'{"name","first_name"}\', ?)', ['"John"']);
-        $builder->from('users')->update(['users.options->name->first_name' => 'John']);
-
-        $builder = $this->getPostgresBuilder();
-        $builder->getConnection()->shouldReceive('update')
-            ->with('update "users" set "options" = jsonb_set("options"::jsonb, \'{"language"}\', \'null\')', []);
-        $builder->from('users')->update(['options->language' => new Raw("'null'")]);
-    }
-
-    public function testPostgresUpdateWrappingJsonArray()
-    {
-        $builder = $this->getPostgresBuilder();
-        $builder->getConnection()->shouldReceive('update')
-            ->with('update "users" set "options" = ?, "meta" = jsonb_set("meta"::jsonb, \'{"tags"}\', ?), "group_id" = 45, "created_at" = ?', [
-                json_encode(['2fa' => false, 'presets' => ['laravel', 'vue']]),
-                json_encode(['white', 'large']),
-                new DateTime('2019-08-06'),
-            ]);
-
-        $builder->from('users')->update([
-            'options' => ['2fa' => false, 'presets' => ['laravel', 'vue']],
-            'meta->tags' => ['white', 'large'],
-            'group_id' => new Raw('45'),
-            'created_at' => new DateTime('2019-08-06'),
-        ]);
-    }
-
-    public function testPostgresUpdateWrappingJsonPathArrayIndex()
-    {
-        $builder = $this->getPostgresBuilder();
-        $builder->getConnection()->shouldReceive('update')
-            ->with('update "users" set "options" = jsonb_set("options"::jsonb, \'{1,"2fa"}\', ?), "meta" = jsonb_set("meta"::jsonb, \'{"tags",0,2}\', ?) where ("options"->1->\'2fa\')::jsonb = \'true\'::jsonb', [
-                'false',
-                '"large"',
-            ]);
-
-        $builder->from('users')->where('options->[1]->2fa', true)->update([
-            'options->[1]->2fa' => false,
-            'meta->tags[0][2]' => 'large',
-        ]);
-    }
-
-    public function testSQLiteUpdateWrappingJsonArray()
-    {
-        $builder = $this->getSQLiteBuilder();
-
-        $builder->getConnection()->shouldReceive('update')
-            ->with('update "users" set "options" = ?, "group_id" = 45, "created_at" = ?', [
-                json_encode(['2fa' => false, 'presets' => ['laravel', 'vue']]),
-                new DateTime('2019-08-06'),
-            ]);
-
-        $builder->from('users')->update([
-            'options' => ['2fa' => false, 'presets' => ['laravel', 'vue']],
-            'group_id' => new Raw('45'),
-            'created_at' => new DateTime('2019-08-06'),
-        ]);
-    }
-
-    public function testSQLiteUpdateWrappingNestedJsonArray()
-    {
-        $builder = $this->getSQLiteBuilder();
-        $builder->getConnection()->shouldReceive('update')
-            ->with('update "users" set "group_id" = 45, "created_at" = ?, "options" = json_patch(ifnull("options", json(\'{}\')), json(?))', [
-                new DateTime('2019-08-06'),
-                json_encode(['name' => 'Taylor', 'security' => ['2fa' => false, 'presets' => ['laravel', 'vue']], 'sharing' => ['twitter' => 'username']]),
-            ]);
-
-        $builder->from('users')->update([
-            'options->name' => 'Taylor',
-            'group_id' => new Raw('45'),
-            'options->security' => ['2fa' => false, 'presets' => ['laravel', 'vue']],
-            'options->sharing->twitter' => 'username',
-            'created_at' => new DateTime('2019-08-06'),
-        ]);
-    }
-
-    public function testSQLiteUpdateWrappingJsonPathArrayIndex()
-    {
-        $builder = $this->getSQLiteBuilder();
-        $builder->getConnection()->shouldReceive('update')
-            ->with('update "users" set "options" = json_patch(ifnull("options", json(\'{}\')), json(?)), "meta" = json_patch(ifnull("meta", json(\'{}\')), json(?)) where json_extract("options", \'$[1]."2fa"\') = true', [
-                '{"[1]":{"2fa":false}}',
-                '{"tags[0][2]":"large"}',
-            ]);
-
-        $builder->from('users')->where('options->[1]->2fa', true)->update([
-            'options->[1]->2fa' => false,
-            'meta->tags[0][2]' => 'large',
-        ]);
-    }
-
-    public function testMySqlWrappingJsonWithString()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->where('items->sku', '=', 'foo-bar');
-        $this->assertSame('select * from `users` where json_unquote(json_extract(`items`, \'$."sku"\')) = ?', $builder->toSql());
-        $this->assertCount(1, $builder->getRawBindings()['where']);
-        $this->assertSame('foo-bar', $builder->getRawBindings()['where'][0]);
-    }
-
-    public function testMySqlWrappingJsonWithInteger()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->where('items->price', '=', 1);
-        $this->assertSame('select * from `users` where json_unquote(json_extract(`items`, \'$."price"\')) = ?', $builder->toSql());
-    }
-
-    public function testMySqlWrappingJsonWithDouble()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->where('items->price', '=', 1.5);
-        $this->assertSame('select * from `users` where json_unquote(json_extract(`items`, \'$."price"\')) = ?', $builder->toSql());
-    }
-
-    public function testMySqlWrappingJsonWithBoolean()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->where('items->available', '=', true);
-        $this->assertSame('select * from `users` where json_extract(`items`, \'$."available"\') = true', $builder->toSql());
-
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->where(new Raw("items->'$.available'"), '=', true);
-        $this->assertSame("select * from `users` where items->'$.available' = true", $builder->toSql());
-    }
-
-    public function testMySqlWrappingJsonWithBooleanAndIntegerThatLooksLikeOne()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->where('items->available', '=', true)->where('items->active', '=', false)->where('items->number_available', '=', 0);
-        $this->assertSame('select * from `users` where json_extract(`items`, \'$."available"\') = true and json_extract(`items`, \'$."active"\') = false and json_unquote(json_extract(`items`, \'$."number_available"\')) = ?', $builder->toSql());
-    }
-
-    public function testJsonPathEscaping()
-    {
-        $expectedWithJsonEscaped = <<<'SQL'
-select json_unquote(json_extract(`json`, '$."''))#"'))
-SQL;
-
-        $builder = $this->getMySqlBuilder();
-        $builder->select("json->'))#");
-        $this->assertEquals($expectedWithJsonEscaped, $builder->toSql());
-
-        $builder = $this->getMySqlBuilder();
-        $builder->select("json->\'))#");
-        $this->assertEquals($expectedWithJsonEscaped, $builder->toSql());
-
-        $builder = $this->getMySqlBuilder();
-        $builder->select("json->\\'))#");
-        $this->assertEquals($expectedWithJsonEscaped, $builder->toSql());
-
-        $builder = $this->getMySqlBuilder();
-        $builder->select("json->\\\'))#");
-        $this->assertEquals($expectedWithJsonEscaped, $builder->toSql());
-    }
-
-    public function testMySqlWrappingJson()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->whereRaw('items->\'$."price"\' = 1');
-        $this->assertSame('select * from `users` where items->\'$."price"\' = 1', $builder->toSql());
-
-        $builder = $this->getMySqlBuilder();
-        $builder->select('items->price')->from('users')->where('users.items->price', '=', 1)->orderBy('items->price');
-        $this->assertSame('select json_unquote(json_extract(`items`, \'$."price"\')) from `users` where json_unquote(json_extract(`users`.`items`, \'$."price"\')) = ? order by json_unquote(json_extract(`items`, \'$."price"\')) asc', $builder->toSql());
-
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->where('items->price->in_usd', '=', 1);
-        $this->assertSame('select * from `users` where json_unquote(json_extract(`items`, \'$."price"."in_usd"\')) = ?', $builder->toSql());
-
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->where('items->price->in_usd', '=', 1)->where('items->age', '=', 2);
-        $this->assertSame('select * from `users` where json_unquote(json_extract(`items`, \'$."price"."in_usd"\')) = ? and json_unquote(json_extract(`items`, \'$."age"\')) = ?', $builder->toSql());
-    }
-
-    public function testPostgresWrappingJson()
-    {
-        $builder = $this->getPostgresBuilder();
-        $builder->select('items->price')->from('users')->where('users.items->price', '=', 1)->orderBy('items->price');
-        $this->assertSame('select "items"->>\'price\' from "users" where "users"."items"->>\'price\' = ? order by "items"->>\'price\' asc', $builder->toSql());
-
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->where('items->price->in_usd', '=', 1);
-        $this->assertSame('select * from "users" where "items"->\'price\'->>\'in_usd\' = ?', $builder->toSql());
-
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->where('items->price->in_usd', '=', 1)->where('items->age', '=', 2);
-        $this->assertSame('select * from "users" where "items"->\'price\'->>\'in_usd\' = ? and "items"->>\'age\' = ?', $builder->toSql());
-
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->where('items->prices->0', '=', 1)->where('items->age', '=', 2);
-        $this->assertSame('select * from "users" where "items"->\'prices\'->>0 = ? and "items"->>\'age\' = ?', $builder->toSql());
-
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->where('items->available', '=', true);
-        $this->assertSame('select * from "users" where ("items"->\'available\')::jsonb = \'true\'::jsonb', $builder->toSql());
-    }
-
-    public function testSqlServerWrappingJson()
-    {
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('items->price')->from('users')->where('users.items->price', '=', 1)->orderBy('items->price');
-        $this->assertSame('select json_value([items], \'$."price"\') from [users] where json_value([users].[items], \'$."price"\') = ? order by json_value([items], \'$."price"\') asc', $builder->toSql());
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->where('items->price->in_usd', '=', 1);
-        $this->assertSame('select * from [users] where json_value([items], \'$."price"."in_usd"\') = ?', $builder->toSql());
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->where('items->price->in_usd', '=', 1)->where('items->age', '=', 2);
-        $this->assertSame('select * from [users] where json_value([items], \'$."price"."in_usd"\') = ? and json_value([items], \'$."age"\') = ?', $builder->toSql());
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->where('items->available', '=', true);
-        $this->assertSame('select * from [users] where json_value([items], \'$."available"\') = \'true\'', $builder->toSql());
-    }
-
-    public function testSqliteWrappingJson()
-    {
-        $builder = $this->getSQLiteBuilder();
-        $builder->select('items->price')->from('users')->where('users.items->price', '=', 1)->orderBy('items->price');
-        $this->assertSame('select json_extract("items", \'$."price"\') from "users" where json_extract("users"."items", \'$."price"\') = ? order by json_extract("items", \'$."price"\') asc', $builder->toSql());
-
-        $builder = $this->getSQLiteBuilder();
-        $builder->select('*')->from('users')->where('items->price->in_usd', '=', 1);
-        $this->assertSame('select * from "users" where json_extract("items", \'$."price"."in_usd"\') = ?', $builder->toSql());
-
-        $builder = $this->getSQLiteBuilder();
-        $builder->select('*')->from('users')->where('items->price->in_usd', '=', 1)->where('items->age', '=', 2);
-        $this->assertSame('select * from "users" where json_extract("items", \'$."price"."in_usd"\') = ? and json_extract("items", \'$."age"\') = ?', $builder->toSql());
-
-        $builder = $this->getSQLiteBuilder();
-        $builder->select('*')->from('users')->where('items->available', '=', true);
-        $this->assertSame('select * from "users" where json_extract("items", \'$."available"\') = true', $builder->toSql());
-    }
-
-    public function testSQLiteOrderBy()
-    {
-        $builder = $this->getSQLiteBuilder();
-        $builder->select('*')->from('users')->orderBy('email', 'desc');
-        $this->assertSame('select * from "users" order by "email" desc', $builder->toSql());
-    }
-
-    public function testSqlServerLimitsAndOffsets()
-    {
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->take(10);
-        $this->assertSame('select top 10 * from [users]', $builder->toSql());
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->skip(10);
-        $this->assertSame('select * from (select *, row_number() over (order by (select 0)) as row_num from [users]) as temp_table where row_num >= 11 order by row_num', $builder->toSql());
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->skip(10)->take(10);
-        $this->assertSame('select * from (select *, row_number() over (order by (select 0)) as row_num from [users]) as temp_table where row_num between 11 and 20 order by row_num', $builder->toSql());
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->skip(11)->take(10)->orderBy('email', 'desc');
-        $this->assertSame('select * from [users] order by [email] desc offset 11 rows fetch next 10 rows only', $builder->toSql());
-
-        $builder = $this->getSqlServerBuilder();
-        $subQueryBuilder = $this->getSqlServerBuilder();
-        $subQuery = function ($query) {
-            return $query->select('created_at')->from('logins')->where('users.name', 'nameBinding')->whereColumn('user_id', 'users.id')->limit(1);
-        };
-        $builder->select('*')->from('users')->where('email', 'emailBinding')->orderBy($subQuery)->skip(10)->take(10);
-        $this->assertSame('select * from [users] where [email] = ? order by (select top 1 [created_at] from [logins] where [users].[name] = ? and [user_id] = [users].[id]) asc offset 10 rows fetch next 10 rows only', $builder->toSql());
-        $this->assertEquals(['emailBinding', 'nameBinding'], $builder->getBindings());
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->take('foo');
-        $this->assertSame('select * from [users]', $builder->toSql());
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->take('foo')->offset('bar');
-        $this->assertSame('select * from [users]', $builder->toSql());
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->offset('bar');
-        $this->assertSame('select * from [users]', $builder->toSql());
-    }
-
-    public function testMySqlSoundsLikeOperator()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->where('name', 'sounds like', 'John Doe');
-        $this->assertSame('select * from `users` where `name` sounds like ?', $builder->toSql());
-        $this->assertEquals(['John Doe'], $builder->getBindings());
-    }
-
     public function testBitwiseOperators()
     {
         $builder = $this->getBuilder();
         $builder->select('*')->from('users')->where('bar', '&', 1);
         $this->assertSame('select * from "users" where "bar" & ?', $builder->toSql());
 
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->where('bar', '#', 1);
-        $this->assertSame('select * from "users" where ("bar" # ?)::bool', $builder->toSql());
-
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->where('range', '>>', '[2022-01-08 00:00:00,2022-01-09 00:00:00)');
-        $this->assertSame('select * from "users" where ("range" >> ?)::bool', $builder->toSql());
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->where('bar', '&', 1);
-        $this->assertSame('select * from [users] where ([bar] & ?) != 0', $builder->toSql());
-
         $builder = $this->getBuilder();
         $builder->select('*')->from('users')->having('bar', '&', 1);
         $this->assertSame('select * from "users" having "bar" & ?', $builder->toSql());
-
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->having('bar', '#', 1);
-        $this->assertSame('select * from "users" having ("bar" # ?)::bool', $builder->toSql());
-
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->having('range', '>>', '[2022-01-08 00:00:00,2022-01-09 00:00:00)');
-        $this->assertSame('select * from "users" having ("range" >> ?)::bool', $builder->toSql());
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->having('bar', '&', 1);
-        $this->assertSame('select * from [users] having ([bar] & ?) != 0', $builder->toSql());
     }
 
     public function testMergeWheresCanMergeWheresAndBindings()
@@ -3552,73 +2150,6 @@ SQL;
         $builder->noValidMethodHere();
     }
 
-    public function testMySqlLock()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('foo')->where('bar', '=', 'baz')->lock();
-        $this->assertSame('select * from `foo` where `bar` = ? for update', $builder->toSql());
-        $this->assertEquals(['baz'], $builder->getBindings());
-
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('foo')->where('bar', '=', 'baz')->lock(false);
-        $this->assertSame('select * from `foo` where `bar` = ? lock in share mode', $builder->toSql());
-        $this->assertEquals(['baz'], $builder->getBindings());
-
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('foo')->where('bar', '=', 'baz')->lock('lock in share mode');
-        $this->assertSame('select * from `foo` where `bar` = ? lock in share mode', $builder->toSql());
-        $this->assertEquals(['baz'], $builder->getBindings());
-    }
-
-    public function testPostgresLock()
-    {
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('foo')->where('bar', '=', 'baz')->lock();
-        $this->assertSame('select * from "foo" where "bar" = ? for update', $builder->toSql());
-        $this->assertEquals(['baz'], $builder->getBindings());
-
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('foo')->where('bar', '=', 'baz')->lock(false);
-        $this->assertSame('select * from "foo" where "bar" = ? for share', $builder->toSql());
-        $this->assertEquals(['baz'], $builder->getBindings());
-
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('foo')->where('bar', '=', 'baz')->lock('for key share');
-        $this->assertSame('select * from "foo" where "bar" = ? for key share', $builder->toSql());
-        $this->assertEquals(['baz'], $builder->getBindings());
-    }
-
-    public function testSqlServerLock()
-    {
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('foo')->where('bar', '=', 'baz')->lock();
-        $this->assertSame('select * from [foo] with(rowlock,updlock,holdlock) where [bar] = ?', $builder->toSql());
-        $this->assertEquals(['baz'], $builder->getBindings());
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('foo')->where('bar', '=', 'baz')->lock(false);
-        $this->assertSame('select * from [foo] with(rowlock,holdlock) where [bar] = ?', $builder->toSql());
-        $this->assertEquals(['baz'], $builder->getBindings());
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('foo')->where('bar', '=', 'baz')->lock('with(holdlock)');
-        $this->assertSame('select * from [foo] with(holdlock) where [bar] = ?', $builder->toSql());
-        $this->assertEquals(['baz'], $builder->getBindings());
-    }
-
-    public function testSelectWithLockUsesWritePdo()
-    {
-        $builder = $this->getMySqlBuilderWithProcessor();
-        $builder->getConnection()->shouldReceive('select')->once()
-            ->with(m::any(), m::any(), false);
-        $builder->select('*')->from('foo')->where('bar', '=', 'baz')->lock()->get();
-
-        $builder = $this->getMySqlBuilderWithProcessor();
-        $builder->getConnection()->shouldReceive('select')->once()
-            ->with(m::any(), m::any(), false);
-        $builder->select('*')->from('foo')->where('bar', '=', 'baz')->lock(false)->get();
-    }
-
     public function testBindingOrder()
     {
         $expectedSql = 'select * from "users" inner join "othertable" on "bar" = ? where "registered" = ? group by "city" having "population" > ? order by match ("foo") against(?)';
@@ -3677,56 +2208,6 @@ SQL;
         $this->assertEquals(['foo', 'bar', 'baz'], $builder->getBindings());
     }
 
-    public function testSubSelect()
-    {
-        $expectedSql = 'select "foo", "bar", (select "baz" from "two" where "subkey" = ?) as "sub" from "one" where "key" = ?';
-        $expectedBindings = ['subval', 'val'];
-
-        $builder = $this->getPostgresBuilder();
-        $builder->from('one')->select(['foo', 'bar'])->where('key', '=', 'val');
-        $builder->selectSub(function ($query) {
-            $query->from('two')->select('baz')->where('subkey', '=', 'subval');
-        }, 'sub');
-        $this->assertEquals($expectedSql, $builder->toSql());
-        $this->assertEquals($expectedBindings, $builder->getBindings());
-
-        $builder = $this->getPostgresBuilder();
-        $builder->from('one')->select(['foo', 'bar'])->where('key', '=', 'val');
-        $subBuilder = $this->getPostgresBuilder();
-        $subBuilder->from('two')->select('baz')->where('subkey', '=', 'subval');
-        $builder->selectSub($subBuilder, 'sub');
-        $this->assertEquals($expectedSql, $builder->toSql());
-        $this->assertEquals($expectedBindings, $builder->getBindings());
-
-        $this->expectException(InvalidArgumentException::class);
-        $builder = $this->getPostgresBuilder();
-        $builder->selectSub(['foo'], 'sub');
-    }
-
-    public function testSubSelectResetBindings()
-    {
-        $builder = $this->getPostgresBuilder();
-        $builder->from('one')->selectSub(function ($query) {
-            $query->from('two')->select('baz')->where('subkey', '=', 'subval');
-        }, 'sub');
-
-        $this->assertSame('select (select "baz" from "two" where "subkey" = ?) as "sub" from "one"', $builder->toSql());
-        $this->assertEquals(['subval'], $builder->getBindings());
-
-        $builder->select('*');
-
-        $this->assertSame('select * from "one"', $builder->toSql());
-        $this->assertEquals([], $builder->getBindings());
-    }
-
-    public function testSqlServerWhereDate()
-    {
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->whereDate('created_at', '=', '2015-09-23');
-        $this->assertSame('select * from [users] where cast([created_at] as date) = ?', $builder->toSql());
-        $this->assertEquals([0 => '2015-09-23'], $builder->getBindings());
-    }
-
     public function testUppercaseLeadingBooleansAreRemoved()
     {
         $builder = $this->getBuilder();
@@ -3746,17 +2227,6 @@ SQL;
         $builder = $this->getBuilder();
         $builder->select('*')->from('users')->where('name', '=', 'Taylor', 'And');
         $this->assertSame('select * from "users" where "name" = ?', $builder->toSql());
-    }
-
-    public function testTableValuedFunctionAsTableInSqlServer()
-    {
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users()');
-        $this->assertSame('select * from [users]()', $builder->toSql());
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users(1,2)');
-        $this->assertSame('select * from [users](1,2)', $builder->toSql());
     }
 
     public function testChunkWithLastChunkComplete()
@@ -4365,351 +2835,6 @@ SQL;
         $builder->select('*')->from('orders')->whereRowValues(['last_update'], '<', [1, 2]);
     }
 
-    public function testWhereJsonContainsMySql()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->whereJsonContains('options', ['en']);
-        $this->assertSame('select * from `users` where json_contains(`options`, ?)', $builder->toSql());
-        $this->assertEquals(['["en"]'], $builder->getBindings());
-
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->whereJsonContains('users.options->languages', ['en']);
-        $this->assertSame('select * from `users` where json_contains(`users`.`options`, ?, \'$."languages"\')', $builder->toSql());
-        $this->assertEquals(['["en"]'], $builder->getBindings());
-
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->where('id', '=', 1)->orWhereJsonContains('options->languages', new Raw("'[\"en\"]'"));
-        $this->assertSame('select * from `users` where `id` = ? or json_contains(`options`, \'["en"]\', \'$."languages"\')', $builder->toSql());
-        $this->assertEquals([1], $builder->getBindings());
-    }
-
-    public function testWhereJsonContainsPostgres()
-    {
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->whereJsonContains('options', ['en']);
-        $this->assertSame('select * from "users" where ("options")::jsonb @> ?', $builder->toSql());
-        $this->assertEquals(['["en"]'], $builder->getBindings());
-
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->whereJsonContains('users.options->languages', ['en']);
-        $this->assertSame('select * from "users" where ("users"."options"->\'languages\')::jsonb @> ?', $builder->toSql());
-        $this->assertEquals(['["en"]'], $builder->getBindings());
-
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->where('id', '=', 1)->orWhereJsonContains('options->languages', new Raw("'[\"en\"]'"));
-        $this->assertSame('select * from "users" where "id" = ? or ("options"->\'languages\')::jsonb @> \'["en"]\'', $builder->toSql());
-        $this->assertEquals([1], $builder->getBindings());
-    }
-
-    public function testWhereJsonContainsSqlite()
-    {
-        $this->expectException(RuntimeException::class);
-
-        $builder = $this->getSQLiteBuilder();
-        $builder->select('*')->from('users')->whereJsonContains('options->languages', ['en'])->toSql();
-    }
-
-    public function testWhereJsonContainsSqlServer()
-    {
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->whereJsonContains('options', true);
-        $this->assertSame('select * from [users] where ? in (select [value] from openjson([options]))', $builder->toSql());
-        $this->assertEquals(['true'], $builder->getBindings());
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->whereJsonContains('users.options->languages', 'en');
-        $this->assertSame('select * from [users] where ? in (select [value] from openjson([users].[options], \'$."languages"\'))', $builder->toSql());
-        $this->assertEquals(['en'], $builder->getBindings());
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->where('id', '=', 1)->orWhereJsonContains('options->languages', new Raw("'en'"));
-        $this->assertSame('select * from [users] where [id] = ? or \'en\' in (select [value] from openjson([options], \'$."languages"\'))', $builder->toSql());
-        $this->assertEquals([1], $builder->getBindings());
-    }
-
-    public function testWhereJsonDoesntContainMySql()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->whereJsonDoesntContain('options->languages', ['en']);
-        $this->assertSame('select * from `users` where not json_contains(`options`, ?, \'$."languages"\')', $builder->toSql());
-        $this->assertEquals(['["en"]'], $builder->getBindings());
-
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->where('id', '=', 1)->orWhereJsonDoesntContain('options->languages', new Raw("'[\"en\"]'"));
-        $this->assertSame('select * from `users` where `id` = ? or not json_contains(`options`, \'["en"]\', \'$."languages"\')', $builder->toSql());
-        $this->assertEquals([1], $builder->getBindings());
-    }
-
-    public function testWhereJsonDoesntContainPostgres()
-    {
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->whereJsonDoesntContain('options->languages', ['en']);
-        $this->assertSame('select * from "users" where not ("options"->\'languages\')::jsonb @> ?', $builder->toSql());
-        $this->assertEquals(['["en"]'], $builder->getBindings());
-
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->where('id', '=', 1)->orWhereJsonDoesntContain('options->languages', new Raw("'[\"en\"]'"));
-        $this->assertSame('select * from "users" where "id" = ? or not ("options"->\'languages\')::jsonb @> \'["en"]\'', $builder->toSql());
-        $this->assertEquals([1], $builder->getBindings());
-    }
-
-    public function testWhereJsonDoesntContainSqlite()
-    {
-        $this->expectException(RuntimeException::class);
-
-        $builder = $this->getSQLiteBuilder();
-        $builder->select('*')->from('users')->whereJsonDoesntContain('options->languages', ['en'])->toSql();
-    }
-
-    public function testWhereJsonDoesntContainSqlServer()
-    {
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->whereJsonDoesntContain('options->languages', 'en');
-        $this->assertSame('select * from [users] where not ? in (select [value] from openjson([options], \'$."languages"\'))', $builder->toSql());
-        $this->assertEquals(['en'], $builder->getBindings());
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->where('id', '=', 1)->orWhereJsonDoesntContain('options->languages', new Raw("'en'"));
-        $this->assertSame('select * from [users] where [id] = ? or not \'en\' in (select [value] from openjson([options], \'$."languages"\'))', $builder->toSql());
-        $this->assertEquals([1], $builder->getBindings());
-    }
-
-    public function testWhereJsonContainsKeyMySql()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->whereJsonContainsKey('users.options->languages');
-        $this->assertSame('select * from `users` where ifnull(json_contains_path(`users`.`options`, \'one\', \'$."languages"\'), 0)', $builder->toSql());
-
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->whereJsonContainsKey('options->language->primary');
-        $this->assertSame('select * from `users` where ifnull(json_contains_path(`options`, \'one\', \'$."language"."primary"\'), 0)', $builder->toSql());
-
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->where('id', '=', 1)->orWhereJsonContainsKey('options->languages');
-        $this->assertSame('select * from `users` where `id` = ? or ifnull(json_contains_path(`options`, \'one\', \'$."languages"\'), 0)', $builder->toSql());
-
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->whereJsonContainsKey('options->languages[0][1]');
-        $this->assertSame('select * from `users` where ifnull(json_contains_path(`options`, \'one\', \'$."languages"[0][1]\'), 0)', $builder->toSql());
-    }
-
-    public function testWhereJsonContainsKeyPostgres()
-    {
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->whereJsonContainsKey('users.options->languages');
-        $this->assertSame('select * from "users" where coalesce(("users"."options")::jsonb ?? \'languages\', false)', $builder->toSql());
-
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->whereJsonContainsKey('options->language->primary');
-        $this->assertSame('select * from "users" where coalesce(("options"->\'language\')::jsonb ?? \'primary\', false)', $builder->toSql());
-
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->where('id', '=', 1)->orWhereJsonContainsKey('options->languages');
-        $this->assertSame('select * from "users" where "id" = ? or coalesce(("options")::jsonb ?? \'languages\', false)', $builder->toSql());
-
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->whereJsonContainsKey('options->languages[0][1]');
-        $this->assertSame('select * from "users" where case when jsonb_typeof(("options"->\'languages\'->0)::jsonb) = \'array\' then jsonb_array_length(("options"->\'languages\'->0)::jsonb) >= 2 else false end', $builder->toSql());
-
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->whereJsonContainsKey('options->languages[-1]');
-        $this->assertSame('select * from "users" where case when jsonb_typeof(("options"->\'languages\')::jsonb) = \'array\' then jsonb_array_length(("options"->\'languages\')::jsonb) >= 1 else false end', $builder->toSql());
-    }
-
-    public function testWhereJsonContainsKeySqlite()
-    {
-        $builder = $this->getSQLiteBuilder();
-        $builder->select('*')->from('users')->whereJsonContainsKey('users.options->languages');
-        $this->assertSame('select * from "users" where json_type("users"."options", \'$."languages"\') is not null', $builder->toSql());
-
-        $builder = $this->getSQLiteBuilder();
-        $builder->select('*')->from('users')->whereJsonContainsKey('options->language->primary');
-        $this->assertSame('select * from "users" where json_type("options", \'$."language"."primary"\') is not null', $builder->toSql());
-
-        $builder = $this->getSQLiteBuilder();
-        $builder->select('*')->from('users')->where('id', '=', 1)->orWhereJsonContainsKey('options->languages');
-        $this->assertSame('select * from "users" where "id" = ? or json_type("options", \'$."languages"\') is not null', $builder->toSql());
-
-        $builder = $this->getSQLiteBuilder();
-        $builder->select('*')->from('users')->whereJsonContainsKey('options->languages[0][1]');
-        $this->assertSame('select * from "users" where json_type("options", \'$."languages"[0][1]\') is not null', $builder->toSql());
-    }
-
-    public function testWhereJsonContainsKeySqlServer()
-    {
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->whereJsonContainsKey('users.options->languages');
-        $this->assertSame('select * from [users] where \'languages\' in (select [key] from openjson([users].[options]))', $builder->toSql());
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->whereJsonContainsKey('options->language->primary');
-        $this->assertSame('select * from [users] where \'primary\' in (select [key] from openjson([options], \'$."language"\'))', $builder->toSql());
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->where('id', '=', 1)->orWhereJsonContainsKey('options->languages');
-        $this->assertSame('select * from [users] where [id] = ? or \'languages\' in (select [key] from openjson([options]))', $builder->toSql());
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->whereJsonContainsKey('options->languages[0][1]');
-        $this->assertSame('select * from [users] where 1 in (select [key] from openjson([options], \'$."languages"[0]\'))', $builder->toSql());
-    }
-
-    public function testWhereJsonDoesntContainKeyMySql()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->whereJsonDoesntContainKey('options->languages');
-        $this->assertSame('select * from `users` where not ifnull(json_contains_path(`options`, \'one\', \'$."languages"\'), 0)', $builder->toSql());
-
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->where('id', '=', 1)->orWhereJsonDoesntContainKey('options->languages');
-        $this->assertSame('select * from `users` where `id` = ? or not ifnull(json_contains_path(`options`, \'one\', \'$."languages"\'), 0)', $builder->toSql());
-
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->whereJsonDoesntContainKey('options->languages[0][1]');
-        $this->assertSame('select * from `users` where not ifnull(json_contains_path(`options`, \'one\', \'$."languages"[0][1]\'), 0)', $builder->toSql());
-    }
-
-    public function testWhereJsonDoesntContainKeyPostgres()
-    {
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->whereJsonDoesntContainKey('options->languages');
-        $this->assertSame('select * from "users" where not coalesce(("options")::jsonb ?? \'languages\', false)', $builder->toSql());
-
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->where('id', '=', 1)->orWhereJsonDoesntContainKey('options->languages');
-        $this->assertSame('select * from "users" where "id" = ? or not coalesce(("options")::jsonb ?? \'languages\', false)', $builder->toSql());
-
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->whereJsonDoesntContainKey('options->languages[0][1]');
-        $this->assertSame('select * from "users" where not case when jsonb_typeof(("options"->\'languages\'->0)::jsonb) = \'array\' then jsonb_array_length(("options"->\'languages\'->0)::jsonb) >= 2 else false end', $builder->toSql());
-
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->whereJsonDoesntContainKey('options->languages[-1]');
-        $this->assertSame('select * from "users" where not case when jsonb_typeof(("options"->\'languages\')::jsonb) = \'array\' then jsonb_array_length(("options"->\'languages\')::jsonb) >= 1 else false end', $builder->toSql());
-    }
-
-    public function testWhereJsonDoesntContainKeySqlite()
-    {
-        $builder = $this->getSQLiteBuilder();
-        $builder->select('*')->from('users')->whereJsonDoesntContainKey('options->languages');
-        $this->assertSame('select * from "users" where not json_type("options", \'$."languages"\') is not null', $builder->toSql());
-
-        $builder = $this->getSQLiteBuilder();
-        $builder->select('*')->from('users')->where('id', '=', 1)->orWhereJsonDoesntContainKey('options->languages');
-        $this->assertSame('select * from "users" where "id" = ? or not json_type("options", \'$."languages"\') is not null', $builder->toSql());
-
-        $builder = $this->getSQLiteBuilder();
-        $builder->select('*')->from('users')->where('id', '=', 1)->orWhereJsonDoesntContainKey('options->languages[0][1]');
-        $this->assertSame('select * from "users" where "id" = ? or not json_type("options", \'$."languages"[0][1]\') is not null', $builder->toSql());
-    }
-
-    public function testWhereJsonDoesntContainKeySqlServer()
-    {
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->whereJsonDoesntContainKey('options->languages');
-        $this->assertSame('select * from [users] where not \'languages\' in (select [key] from openjson([options]))', $builder->toSql());
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->where('id', '=', 1)->orWhereJsonDoesntContainKey('options->languages');
-        $this->assertSame('select * from [users] where [id] = ? or not \'languages\' in (select [key] from openjson([options]))', $builder->toSql());
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->where('id', '=', 1)->orWhereJsonDoesntContainKey('options->languages[0][1]');
-        $this->assertSame('select * from [users] where [id] = ? or not 1 in (select [key] from openjson([options], \'$."languages"[0]\'))', $builder->toSql());
-    }
-
-    public function testWhereJsonLengthMySql()
-    {
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->whereJsonLength('options', 0);
-        $this->assertSame('select * from `users` where json_length(`options`) = ?', $builder->toSql());
-        $this->assertEquals([0], $builder->getBindings());
-
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->whereJsonLength('users.options->languages', '>', 0);
-        $this->assertSame('select * from `users` where json_length(`users`.`options`, \'$."languages"\') > ?', $builder->toSql());
-        $this->assertEquals([0], $builder->getBindings());
-
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->where('id', '=', 1)->orWhereJsonLength('options->languages', new Raw('0'));
-        $this->assertSame('select * from `users` where `id` = ? or json_length(`options`, \'$."languages"\') = 0', $builder->toSql());
-        $this->assertEquals([1], $builder->getBindings());
-
-        $builder = $this->getMySqlBuilder();
-        $builder->select('*')->from('users')->where('id', '=', 1)->orWhereJsonLength('options->languages', '>', new Raw('0'));
-        $this->assertSame('select * from `users` where `id` = ? or json_length(`options`, \'$."languages"\') > 0', $builder->toSql());
-        $this->assertEquals([1], $builder->getBindings());
-    }
-
-    public function testWhereJsonLengthPostgres()
-    {
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->whereJsonLength('options', 0);
-        $this->assertSame('select * from "users" where jsonb_array_length(("options")::jsonb) = ?', $builder->toSql());
-        $this->assertEquals([0], $builder->getBindings());
-
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->whereJsonLength('users.options->languages', '>', 0);
-        $this->assertSame('select * from "users" where jsonb_array_length(("users"."options"->\'languages\')::jsonb) > ?', $builder->toSql());
-        $this->assertEquals([0], $builder->getBindings());
-
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->where('id', '=', 1)->orWhereJsonLength('options->languages', new Raw('0'));
-        $this->assertSame('select * from "users" where "id" = ? or jsonb_array_length(("options"->\'languages\')::jsonb) = 0', $builder->toSql());
-        $this->assertEquals([1], $builder->getBindings());
-
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->where('id', '=', 1)->orWhereJsonLength('options->languages', '>', new Raw('0'));
-        $this->assertSame('select * from "users" where "id" = ? or jsonb_array_length(("options"->\'languages\')::jsonb) > 0', $builder->toSql());
-        $this->assertEquals([1], $builder->getBindings());
-    }
-
-    public function testWhereJsonLengthSqlite()
-    {
-        $builder = $this->getSQLiteBuilder();
-        $builder->select('*')->from('users')->whereJsonLength('options', 0);
-        $this->assertSame('select * from "users" where json_array_length("options") = ?', $builder->toSql());
-        $this->assertEquals([0], $builder->getBindings());
-
-        $builder = $this->getSQLiteBuilder();
-        $builder->select('*')->from('users')->whereJsonLength('users.options->languages', '>', 0);
-        $this->assertSame('select * from "users" where json_array_length("users"."options", \'$."languages"\') > ?', $builder->toSql());
-        $this->assertEquals([0], $builder->getBindings());
-
-        $builder = $this->getSQLiteBuilder();
-        $builder->select('*')->from('users')->where('id', '=', 1)->orWhereJsonLength('options->languages', new Raw('0'));
-        $this->assertSame('select * from "users" where "id" = ? or json_array_length("options", \'$."languages"\') = 0', $builder->toSql());
-        $this->assertEquals([1], $builder->getBindings());
-
-        $builder = $this->getSQLiteBuilder();
-        $builder->select('*')->from('users')->where('id', '=', 1)->orWhereJsonLength('options->languages', '>', new Raw('0'));
-        $this->assertSame('select * from "users" where "id" = ? or json_array_length("options", \'$."languages"\') > 0', $builder->toSql());
-        $this->assertEquals([1], $builder->getBindings());
-    }
-
-    public function testWhereJsonLengthSqlServer()
-    {
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->whereJsonLength('options', 0);
-        $this->assertSame('select * from [users] where (select count(*) from openjson([options])) = ?', $builder->toSql());
-        $this->assertEquals([0], $builder->getBindings());
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->whereJsonLength('users.options->languages', '>', 0);
-        $this->assertSame('select * from [users] where (select count(*) from openjson([users].[options], \'$."languages"\')) > ?', $builder->toSql());
-        $this->assertEquals([0], $builder->getBindings());
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->where('id', '=', 1)->orWhereJsonLength('options->languages', new Raw('0'));
-        $this->assertSame('select * from [users] where [id] = ? or (select count(*) from openjson([options], \'$."languages"\')) = 0', $builder->toSql());
-        $this->assertEquals([1], $builder->getBindings());
-
-        $builder = $this->getSqlServerBuilder();
-        $builder->select('*')->from('users')->where('id', '=', 1)->orWhereJsonLength('options->languages', '>', new Raw('0'));
-        $this->assertSame('select * from [users] where [id] = ? or (select count(*) from openjson([options], \'$."languages"\')) > 0', $builder->toSql());
-        $this->assertEquals([1], $builder->getBindings());
-    }
-
     public function testFromSub()
     {
         $builder = $this->getBuilder();
@@ -4755,34 +2880,12 @@ SQL;
         $this->assertSame('select * from (select max(last_seen_at) as last_seen_at from "user_sessions") as "sessions"', $builder->toSql());
     }
 
-    public function testFromRawOnSqlServer()
-    {
-        $builder = $this->getSqlServerBuilder();
-        $builder->fromRaw('dbo.[SomeNameWithRoundBrackets (test)]');
-        $this->assertSame('select * from dbo.[SomeNameWithRoundBrackets (test)]', $builder->toSql());
-    }
-
     public function testFromRawWithWhereOnTheMainQuery()
     {
         $builder = $this->getBuilder();
         $builder->fromRaw(new Raw('(select max(last_seen_at) as last_seen_at from "sessions") as "last_seen_at"'))->where('last_seen_at', '>', '1520652582');
         $this->assertSame('select * from (select max(last_seen_at) as last_seen_at from "sessions") as "last_seen_at" where "last_seen_at" > ?', $builder->toSql());
         $this->assertEquals(['1520652582'], $builder->getBindings());
-    }
-
-    public function testFromQuestionMarkOperatorOnPostgres()
-    {
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->where('roles', '?', 'superuser');
-        $this->assertSame('select * from "users" where "roles" ?? ?', $builder->toSql());
-
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->where('roles', '?|', 'superuser');
-        $this->assertSame('select * from "users" where "roles" ??| ?', $builder->toSql());
-
-        $builder = $this->getPostgresBuilder();
-        $builder->select('*')->from('users')->where('roles', '?&', 'superuser');
-        $this->assertSame('select * from "users" where "roles" ??& ?', $builder->toSql());
     }
 
     public function testClone()
@@ -4833,54 +2936,6 @@ SQL;
         $processor = m::mock(Processor::class);
 
         return new Builder($this->getConnection(), $grammar, $processor);
-    }
-
-    protected function getPostgresBuilder()
-    {
-        $grammar = new PostgresGrammar;
-        $processor = m::mock(Processor::class);
-
-        return new Builder($this->getConnection(), $grammar, $processor);
-    }
-
-    protected function getMySqlBuilder()
-    {
-        $grammar = new MySqlGrammar;
-        $processor = m::mock(Processor::class);
-
-        return new Builder(m::mock(ConnectionInterface::class), $grammar, $processor);
-    }
-
-    protected function getSQLiteBuilder()
-    {
-        $grammar = new SQLiteGrammar;
-        $processor = m::mock(Processor::class);
-
-        return new Builder(m::mock(ConnectionInterface::class), $grammar, $processor);
-    }
-
-    protected function getSqlServerBuilder()
-    {
-        $grammar = new SqlServerGrammar;
-        $processor = m::mock(Processor::class);
-
-        return new Builder($this->getConnection(), $grammar, $processor);
-    }
-
-    protected function getMySqlBuilderWithProcessor()
-    {
-        $grammar = new MySqlGrammar;
-        $processor = new MySqlProcessor;
-
-        return new Builder(m::mock(ConnectionInterface::class), $grammar, $processor);
-    }
-
-    protected function getPostgresBuilderWithProcessor()
-    {
-        $grammar = new PostgresGrammar;
-        $processor = new PostgresProcessor;
-
-        return new Builder(m::mock(ConnectionInterface::class), $grammar, $processor);
     }
 
     /**
